Convert anndata to Giotto

Converts a spatial anndata (e.g. scanpy) .h5ad file into a Giotto object

Usage

anndataToGiotto(
  anndata_path = NULL,
  n_key_added = NULL,
  spatial_n_key_added = NULL,
  deluanay_spat_net = TRUE,
  spat_unit = NULL,
  feat_type = NULL,
  h5_file = NULL,
  python_path = NULL,
  env_name = "giotto_env"
)

Arguments

anndata_path

path to the .h5ad file

n_key_added

equivalent of "key_added" argument from scanpy.pp.neighbors(). If multiple spatial networks are in the anndata object, a list of key_added terms may be provided. If converting an anndata object from giottoToAnnData, a .txt file may be provided, which was generated in that function, i.e. spat_unit_feat_type_nn_network_keys_added.txt Cannot be "spatial". This becomes the name of the nearest network in the gobject.

spatial_n_key_added

equivalent of "key_added" argument from squidpy.gr.spatial_neighbors. If multiple spatial networks are in the anndata object, a list of key_added terms may be provided. If converting an anndata object from giottoToAnnData, a .txt file may be provided, which was generated in that function, i.e. spat_unit_feat_type_spatial_network_keys_added.txt Cannot be the same as n_key_added.

deluanay_spat_net

binary parameter for spatial network. If TRUE, the spatial network is a deluanay network.

spat_unit

desired spatial unit for conversion, default NULL

feat_type

desired feature type for conversion, default NULL

h5_file

name to create and on-disk HDF5 file

python_path

path to python executable within a conda/miniconda environment

env_name

name of environment containing python_path executable

Value

Giotto object

Details

Function in beta. Converts a .h5ad file into a Giotto object. The returned Giotto Object will take default insructions with the exception of the python path, which may be customized. See changeGiottoInstructions to modify instructions after creation.