Coerce to data.table

Coerce to data.table if possible

Usage

# S3 method for class 'SpatVector'
as.data.table(
  x,
  keep.rownames = FALSE,
  geom = NULL,
  include_values = TRUE,
  ...
)

# S3 method for class 'giottoPolygon'
as.data.table(x, ...)

# S3 method for class 'giottoPoints'
as.data.table(x, ...)

Arguments

x

The object to coerce

keep.rownames

This argument is ignored

geom

character or NULL. If not NULL, either "XY", "WKT", or "HEX", to get the geometry included in coordinates of each point or vertex, Well-Known-Text or hexadecimal notation.

include_values

whether to include attributes information when geom is 'XY'

...

additional arguments to pass

See also