createGiottoLargeImageList

Creates a list of large giotto images that can be added to a Giotto object. Generates deep copy of SpatRaster

Usage

createGiottoLargeImageList(
  raster_objects,
  names = "image",
  negative_y = TRUE,
  extent = NULL,
  use_rast_ext = FALSE,
  image_transformations = NULL,
  flip_vertical = FALSE,
  flip_horizontal = FALSE,
  xmax_bound = NULL,
  xmin_bound = NULL,
  ymax_bound = NULL,
  ymin_bound = NULL,
  scale_factor = 1,
  verbose = TRUE
)

Arguments

raster_objects

vector of image paths or terra SpatRaster image objects

names

vector of names for the images

negative_y

Map image to negative y spatial values if TRUE. Meaning that origin is in upper left instead of lower left.

extent

SpatExtent object to assign spatial extent. Takes priority unless use_rast_ext is TRUE.

use_rast_ext

Use extent from input raster object

image_transformations

vector of sequential image transformations - under construction

flip_vertical

flip raster in a vertical manner

flip_horizontal

flip raster in a horizontal manner

xmax_bound, xmin_bound, ymax_bound, ymin_bound

assign min and max x and y values for image spatial placement

scale_factor

scaling of image dimensions relative to spatial coordinates

verbose

be verbose

Value

a list with giottoLargeImage objects

Details

See createGiottoLargeImage

Examples

image_test <- system.file("extdata/toy_intensity.tif",
    package = "GiottoClass"
)

createGiottoLargeImageList(raster_objects = image_test)
#> [[1]]
#> An object of class giottoLargeImage : "image"
#> Image extent            : 0, 1091, -1107, 0 (xmin, xmax, ymin, ymax)
#> Original image extent   : 0, 1091, -1107, 0 (xmin, xmax, ymin, ymax)
#> Scale factor            : 1, 1 (x, y)
#> Resolution              : 1, 1 (x, y)
#> Layers                  : 1 
#> Name                    : mini_dataset_dapi_z0 
#> Estimated max intensity : 168 
#> Estimated min intensity : 0 
#> Values                  : integers
#> File path               : '/home/runner/work/_temp/Library/GiottoClass/extdata/toy_intensity.tif'
#>