Create giotto polygons object

Create a giottoPolygon object that is used to represent spatial annotations and polygons. Inputs can be from a structured data.frame object where three of the columns should correspond to x/y vertices and the polygon ID and additional columns are set as attributes, a spatial file such as wkt, .shp, or .GeoJSON, or a mask file (e.g. segmentation results)

Usage

# S4 method for class 'SpatVector'
createGiottoPolygon(x, name = "cell", calc_centroids = FALSE, verbose = TRUE)

# S4 method for class 'SpatRaster'
createGiottoPolygon(
  x,
  name = "cell",
  mask_method = c("guess", "single", "multiple"),
  remove_background_polygon = FALSE,
  background_algo = c("range"),
  fill_holes = TRUE,
  poly_IDs = NULL,
  flip_vertical = TRUE,
  shift_vertical_step = TRUE,
  flip_horizontal = TRUE,
  shift_horizontal_step = TRUE,
  fix_multipart = TRUE,
  remove_unvalid_polygons = TRUE,
  calc_centroids = FALSE,
  verbose = TRUE
)

# S4 method for class 'data.frame'
createGiottoPolygon(
  x,
  name = "cell",
  calc_centroids = FALSE,
  skip_eval_dfr = FALSE,
  copy_dt = TRUE,
  verbose = TRUE
)

# S4 method for class 'character'
createGiottoPolygon(x, ...)

createGiottoPolygonsFromMask(
  maskfile,
  mask_method = c("guess", "single", "multiple"),
  name = "cell",
  remove_background_polygon = FALSE,
  background_algo = c("range"),
  fill_holes = TRUE,
  poly_IDs = NULL,
  flip_vertical = TRUE,
  shift_vertical_step = TRUE,
  flip_horizontal = TRUE,
  shift_horizontal_step = TRUE,
  calc_centroids = FALSE,
  fix_multipart = TRUE,
  remove_unvalid_polygons = TRUE
)

createGiottoPolygonsFromDfr(
  segmdfr,
  name = "cell",
  calc_centroids = FALSE,
  verbose = TRUE,
  skip_eval_dfr = FALSE,
  copy_dt = TRUE
)

createGiottoPolygonsFromGeoJSON(
  GeoJSON,
  name = "cell",
  calc_centroids = FALSE,
  verbose = TRUE
)

Arguments

x

input. Filepath to a .GeoJSON or a mask image file. Can also be a data.frame with vertex 'x', 'y', and 'poly_ID' information.

name

name for the giottoPolygon object created

calc_centroids

(default FALSE) calculate centroids for polygons

verbose

be verbose

mask_method

how the mask file defines individual segmentation annotations

remove_background_polygon

try to remove background polygon (default: FALSE)

background_algo

algorithm to remove background polygon

fill_holes

fill holes within created polygons

poly_IDs

unique names for each polygon in the mask file

flip_vertical

flip mask figure in a vertical manner

shift_vertical_step

shift vertical (boolean or numerical)

flip_horizontal

flip mask figure in a horizontal manner

shift_horizontal_step

shift horizontal (boolean or numerical)

fix_multipart

try to split polygons with multiple parts (default: TRUE)

remove_unvalid_polygons

remove unvalid polygons (default: TRUE)

skip_eval_dfr

(default FALSE) skip evaluation of provided dataframe

copy_dt

(default TRUE) if segmdfr is provided as dt, this determines whether a copy is made

...

additional params to pass. For character method, params pass to SpatRaster or SpatVector methods, depending on whether x was a filepath to a maskfile or a spatial file (ex: wkt, shp, GeoJSON) respectively.

maskfile

path to mask file

segmdfr

data.frame-like object with polygon coordinate information (x, y, poly_ID) with x and y being vertex information for the polygon referenced by poly_ID. See details for how columns are selected for coordinate and ID information.

GeoJSON

path to .GeoJSON file

Value

a giotto polygon object

Details

When determining which column within the tabular data is intended to provide polygon information, Giotto first checks the column names for 'x', 'y', and 'poly_ID'. If any of these are discovered, they are directly selected. If this is not discovered then Giotto checks the data type of the columns and selects the first 'character' type column to be 'poly_ID' and the first two 'numeric' columns as 'x' and 'y' respectively. If this is also unsuccessful then poly_ID defaults to the 3rd column. 'x' and 'y' then default to the 1st and 2nd columns.