.plot_giottolargeimage

Plot a downsampled version of giottoLargeImage. Cropping can increase plot resolution of region of interest.

Usage

.plot_giottolargeimage(
  gobject = NULL,
  largeImage_name = NULL,
  giottoLargeImage = NULL,
  crop_extent = NULL,
  xmax_crop = NULL,
  xmin_crop = NULL,
  ymax_crop = NULL,
  ymin_crop = NULL,
  max_intensity = NULL,
  asRGB = FALSE,
  stretch = NULL,
  axes = TRUE,
  smooth = TRUE,
  mar = c(3, 5, 1.5, 1),
  legend = FALSE,
  maxcell = 5e+05,
  col = grDevices::grey.colors(n = 256, start = 0, end = 1, gamma = 1),
  asp = 1,
  ...
)

Arguments

gobject

giotto object

largeImage_name

name of giottoLargeImage

giottoLargeImage

giottoLargeImage object

crop_extent

(optional) extent object to focus on specific region of image

xmax_crop, xmin_crop, ymax_crop, ymin_crop

(optional) crop min/max x and y bounds

max_intensity

(optional) value to treat as maximum intensity in color scale

asRGB

(optional) logical. Force RGB plotting if not automatically detected

stretch

character. Option to stretch the values to increase contrast: "lin" linear or "hist" (histogram)

axes

boolean. Default = TRUE. Whether to draw axes

smooth

boolean. default = TRUE. whether to apply smoothing on the image

mar

plot margins default = c(3,5,1.5,1)

legend

whether to plot legend of color scale (grayscale only). default = FALSE

maxcell

positive integer. Maximum number of image cells to use for the plot

col

character. Colors for single channel images. The default is grDevices::grey.colors(n = 256, start = 0, end = 1, gamma = 1). It can also be a data.frame with two columns (value, color) to get a "classes" type legend or with three columns (from, to, color) to get an "interval" type legend

asp

numeric. (default = 1) specific aspect ratio to use

...

additional params to pass to terra::plot or terra::plotRGB depending depending on image type

Value

plot