loadGiotto

Saves a Giotto object to a specific folder structure

Usage

loadGiotto(
  path_to_folder,
  load_params = list(),
  reconnect_giottoImage = TRUE,
  python_path = NULL,
  verbose = TRUE
)

Arguments

path_to_folder

path to folder where Giotto object was stored with saveGiotto

load_params

additional parameters for loading or reading giotto object

reconnect_giottoImage

(default = TRUE) whether to attempt reconnection of magick based image objects

python_path

(optional) manually set your python path

verbose

be verbose

Value

Giotto object

Details

Works together with saveGiotto to save and re-load Giotto objects. Additional load_params need to be provided as a list and will go to readRDS or qread You can set the python path, alternatively it will look for an existing Giotto python environment.