Polygon and Point Relationships

Utility class for storing overlaps relationships between polygons and points in a sparse data.table format. Retrieve the unique ID index of overlapped points [i, ]. Get indices of which polys are overlapping specific feature species using [, j].

Subsetting with ids = FALSE and [i, j] indexing is also supported.

Supports as.matrix for conversion to dgCMatrix. Contained poly and feature names simplify rownames/colnames and empty row/col creation.

Usage

# S4 method for class 'overlapPointDT,gIndex,missing,missing'
x[i, j, ..., use_names = FALSE, ids = TRUE, drop]

# S4 method for class 'overlapPointDT,missing,gIndex,missing'
x[i, j, ..., use_names = FALSE, ids = TRUE, drop]

# S4 method for class 'overlapPointDT,gIndex,gIndex,missing'
x[i, j, ..., use_names = FALSE, drop]

Arguments

x

object

i

numeric, character, logical. Index of or name of poly in overlapping polygons

j

numeric, character, logical. Index of or name of feature being overlapped.

...

additional params to pass (none implemented)

use_names

logical (default = FALSE). Whether to return as integer indices or with character ids.

ids

logical (default = TRUE). Whether to return the requested integer indices (TRUE) or the subset overlap object (FALSE).

drop

not used.

Value

integer or character if only i or j provided, depending on use_names. A subset overlapPointDT if both i and j are used.

Slots

data

data.table. Table containing 3 integer cols:

poly - polygon index. Maps to spat_ids slot.

feat - feat_ID_uniq (unique integer identifier) of a point detection

feat_id_index - index of feature name mapping in @feat_ids slot.

spat_unit

character. Spatial unit (usually name of polygons information)

feat_type

character. Feature type (usually name of points information)

provenance

character. provenance information

spat_ids

character. Polygon names

feat_ids

character. Feature names

nfeats

integer (optional metadata). How many feature points were used in overlap operation. Gives an idea of sparsity, but has no effect on processing.

Examples

g <- GiottoData::loadGiottoMini("vizgen")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 4. read Giotto image information
#> python already initialized in this session
#>  active environment : '/usr/bin/python3'
#>  python version : 3.12
poly <- g[["spatial_info", "z0"]][[1]]
ovlp <- overlaps(poly, "rna")
ovlp
#> <overlapPointDT>
#> spat_unit : "z0"
#> feat_type : "rna"
#> provenance: z0 
#> * polygons : 498
#> * features : 35889
#> * relations: 15050

as.matrix(ovlp)
#> 336 x 498 sparse Matrix of class "dgCMatrix"
#>   [[ suppressing 33 column names ‘40951783403982682273285375368232495429’, ‘240649020551054330404932383065726870513’, ‘274176126496863898679934791272921588227’ ... ]]
#>                                                                               
#> Mlc1      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gprc5b    . . 1 . 1 . . .  1  . 2 . .  . . . . 1 .  .  . .  . . . . 1 .  . . 1
#> Gfap      . . . 1 1 . . . 27  . . . .  . . . . . 2  .  . 2  . 1 . . 2 3  . . 6
#> Ednrb     . . 3 . . 1 . .  .  . . 1 .  . . . . . 1  .  . .  . . . . . .  . . .
#> Sox9      . . . 1 . . . .  3  . . . .  . . . 1 . 3  .  . .  . . . . 1 .  . . .
#> Aqp4      . . . 1 1 1 . .  3  . 2 1 .  . . . . . 8  .  . .  . . . . 2 .  . . 4
#> Gjc3      . . . . . . . .  4  . . . .  . . . . . .  .  . .  . . . . . .  . . 1
#> Sox8      . . 3 . . . . .  1  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ntsr2     . . 7 1 2 . . .  .  . 6 7 .  . . . . . 5  .  . .  . . . . . .  . . 2
#> Adcyap1r1 . . . 1 . . . .  .  . . . .  . . . . . 2  .  . .  . . . . . .  . . 1
#> Smo       . . 1 . . . . .  .  . . . .  . 1 . 1 . .  .  2 .  1 . . . . .  1 . .
#> Gpr161    . . . . . . . .  .  . 1 . .  . . . . . .  .  . .  . . . . . .  . . .
#> C1qb      . . . . 6 . . .  .  . . . .  . . . . . .  .  . .  . . . . . 1  . . .
#> Olig1     . . . 1 . . . .  4  . . . .  . . . . . .  .  . .  . 1 . . 1 .  . . .
#> C1qa      . . . 1 9 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Lmtk2     . . . . . . 1 .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Pth1r     . . . . . . . .  .  . 1 . .  . . . . . .  .  . .  . . 1 . . .  . . .
#> Ephb6     . . . . . . . .  .  . . . .  . . . . . 1  .  . .  . . . . . .  . . 2
#> Gpr26     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gabbr1    . . . . . . . .  .  . 1 2 .  . . . . . .  1  . .  . 1 . . . .  . . .
#> Adra2a    . . . . . . . .  .  . . . .  . . . . . 1  .  . .  . . . . . .  . . .
#> Gabbr2    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Timp4     . . 2 . . . . .  1  . 1 . .  . . . . . .  .  . .  . . . . . .  . . .
#> Abcc9     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . 3 . . .  . . .
#> Csf1r     . . . . 1 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ntrk2     . . . 2 2 . . .  .  . 2 1 .  . . . . . 3  .  . 1  . . . . . .  . . 1
#> Adgrg1    . . 1 . 1 . . .  1  . 3 3 .  . . . . . .  .  . .  . . . . . .  . . 1
#> Ntrk3     . . . . . . 1 .  .  . . . .  . . . . . 1  1  . .  . . . . . .  1 . .
#> Ddr1      . . . . . . . .  1  . 1 . .  . 1 . 1 . .  1  . .  . . . . . .  2 . .
#> Erbb4     . . . . . . . .  1  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Selplg    . . . 1 2 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Cspg5     . . . . . 1 . .  .  . 2 1 .  . . . . . 2  .  . .  . 1 . . . .  . . .
#> Dlk1      . . . . . . . .  .  . . . .  . . . . 1 .  .  . .  . . . . . .  . . .
#> Gpr37l1   . . 1 . 2 . . .  1  . 4 5 .  . . . . . 7  .  . .  . . . . 1 .  . . 1
#> Fzd1      . . . . . . . .  .  . . . .  1 . . . . .  .  . .  . . . . . .  . . .
#> Fgfr3     . . 1 1 1 . . .  .  . 2 3 .  . . . . . 3  .  . .  . . . . . .  . . 1
#> Slc25a18  . . 1 . 1 . . .  .  . 1 . .  . . . . . .  .  . .  . . 1 . . .  . . .
#> Igf1r     . . . . . . . .  .  . . 1 .  1 . . . 1 .  .  1 .  1 . . . . .  1 . .
#> Adgrl1    . . . . . . . .  .  . 1 1 .  . . . . . .  .  . .  1 . . . . .  . . .
#> S1pr1     . . . . . . . .  .  1 2 2 .  1 . . . . 2  1  . .  . . . . . .  . . .
#> Ryk       . . . . 1 . 1 .  1  . . . .  . . . . . .  .  . .  2 . 1 . . .  1 . .
#> Grm5      . . . . . . . .  .  . . . .  . . . . . .  .  . 2  . . . . . .  . . .
#> Eomes     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Axl       . . 1 . . . . .  1  . 2 2 .  1 . . . . 1  1  . 1  . . . . . .  1 . 1
#> Efemp1    . . . . . . 2 .  .  1 . . .  4 . . 2 . 1  .  3 .  8 . . . . 1  2 . .
#> Adgrb3    . . . . . . . .  .  . . 3 .  . . . . . .  .  . .  . . . . . .  . . 2
#> Sstr4     . . . . . . . .  .  . . . .  . . . . . .  .  1 .  3 . . . . .  2 . .
#> Fn1       . . . . . . . .  . 10 . . .  7 3 . 9 1 .  .  1 . 16 . . . . .  2 . .
#> Adora1    . . . . . . . .  .  . 1 . .  . . . . . .  .  . .  . . . . . .  . . .
#> Erbb2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gper1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . 1 . . .  . . .
#> Gramd3    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . 1
#> Mertk     . . 1 . 3 . . .  .  . . 1 .  . . . . 1 1  .  . .  . . . . . .  . . .
#> Slc32a1   . . . . 1 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adgra1    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adgrl2    . . . . . . 1 .  .  . . . .  1 . . . . .  .  . .  1 . . . . .  . . .
#> Ptgfr     . . . . . . . .  .  . . . .  . . . . 1 .  .  . .  1 . . . . .  . . .
#> Fzd2      . . . . . . 1 .  .  . . . .  . 1 . 2 . .  .  . .  3 . . . . .  3 . .
#> Aldh1l1   . . 1 . . 1 . .  .  . . 1 .  . 1 . . . 1  .  . .  1 . . . . .  . . .
#> Adgrg6    . . . . . . . .  .  2 . . .  . . . . . .  .  1 .  . . . . . .  . . .
#> Egfr      . . . 1 . . 1 .  .  . . . .  1 1 . 1 . .  .  2 .  1 . 1 . . .  . . .
#> Gpr162    . . . . . . . .  .  . . . .  . . . . . 1  .  . .  . . . . . .  1 . .
#> Lpar4     . . . . . . . .  .  1 . . .  . . . . . .  .  . .  . . . . . .  1 . .
#> Kit       . . . . . . . .  .  1 . . .  . . . . . .  1  . .  . . . . . .  . . .
#> Ptgdr     . . . . . . . .  .  1 . . .  1 1 . 2 . .  .  . .  . . . . . .  2 . .
#> Gad1      . . . . . . . .  .  . 1 1 .  . . . . . .  .  . .  . . . . . .  . . .
#> Lgr4      . . . . . . . .  1  . 1 . .  1 . . . 1 1  1  . .  . . . . . .  . . .
#> Tyro3     . . . . . . . .  2  . . . .  . . . . . .  .  . .  . . . . . .  . . 1
#> Atp13a5   . . . . . 1 . .  .  . . . .  . . 2 1 . .  .  . .  1 . 3 . . .  . . .
#> Ddr2      . . . . . . . .  .  1 . . .  3 1 . . . .  1  . .  4 . . . . .  . . .
#> Rspo3     . . . . . . . .  .  . . . .  3 . 1 2 . .  3  1 .  3 . . . . .  . . .
#> Arhgap29  . . . . . 2 . .  .  1 . . .  1 . 2 1 . .  .  . .  2 . . . 1 .  . . .
#> Ackr3     . . . . . . . .  .  1 . . .  . . 2 . . .  .  . .  . . . . . .  . . .
#> Gpr146    . . . . 1 1 1 .  .  . . 1 .  . 1 . . . .  .  . .  . . . . . .  . . .
#> Fzd7      . . . . . 1 . .  .  . . . .  . . 1 . . .  .  1 .  2 . . . . .  . . .
#> Slc47a1   . . . . . . 7 .  .  . . . . 12 2 . 5 . . 10 10 .  5 . . . . . 12 . .
#> Insr      . . . . . . . .  .  . . 1 .  . . . 1 . .  .  . .  . . . . . .  . . .
#> Ptger3    . . . . . . . .  .  . . . .  . . . . . .  1  . .  . . . . . .  . . .
#> Bdkrb1    . . . . . . 1 .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Traf4     . . . . . . 1 .  .  . . . .  1 . . 1 . .  .  . .  . . . . . .  . . .
#> Celsr2    . . 2 . . . . .  1  . . . .  . . . . . 1  1  . .  . . . . . 1  . . .
#> Fzd4      . . . . . . 1 .  .  . . . .  . . . 3 . .  1  . .  . . . . . .  . . .
#> Fgfr2     . . . . . . . .  1  . . 1 .  2 . 1 . . 1  1  . .  1 . . . . .  . . .
#> Cx3cr1    . . . . 4 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Anxa11    . . . . . . . .  .  . . . .  1 . 1 . . .  .  . .  . . . . . .  . . .
#> Adgrb1    . . . . 1 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . 1
#> Lpar1     . . . . . . . .  .  . . . .  1 . . . 4 .  .  . .  1 . . . . .  . . .
#> Gpr17     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . 3  . . .
#> Gpr173    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ror1      . . . . . . . .  .  . . . .  . . . . . .  1  . .  . . . . . .  . . .
#> Cxcl12    . . . . . 1 1 .  .  . . . .  5 . . . . .  .  1 .  3 . . . . .  . . .
#> Kiss1r    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Epha4     . . . . . 1 . .  .  . . . .  . . . . . .  .  . .  1 . . . . .  . . .
#> Gpr182    . . . . . . . .  .  . . . .  . . . . . .  .  . .  1 . 1 . . .  . . .
#> Man1a     . . . . . . . .  .  . . . .  2 . . . . 1  1  . .  . . . . . .  . . .
#> Gpr4      . . . . . . 1 .  .  . . . .  1 . . 1 1 1  .  1 .  . . 1 . . .  1 . .
#> P2ry14    . . . . . . . .  .  . . . .  . . . 1 . .  .  . .  . . 1 . . .  . . .
#> Hcar1     . . . . . . . .  .  . . . .  . . . 2 . .  .  . .  2 . . . . .  2 . .
#> Ephb1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adrb1     . . . . 1 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Grm3      . . . . . . . .  1  . 1 2 .  . . . . . 1  .  . .  . . . . . .  . . .
#> P2ry1     . . . . . . . .  .  . . . .  . 2 1 . . .  .  . .  . . . . . .  . . .
#> Nrp2      . . . . . . . .  .  . . . .  . . . . . .  .  . .  1 . . . . .  . . .
#> Ptk7      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Slco1a4   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Tmem108   . . . . . . . .  1  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Hrh3      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Insrr     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Cmklr1    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Opn3      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adra1b    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr62     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Npy1r     . . . . . . . .  .  . . . .  . . 1 . . .  .  1 .  . . . . . .  . . .
#> Nrp1      . . 1 . . . . .  .  . 1 . .  1 . . . . .  1  . .  1 . . . . .  2 . .
#> Fzd6      . . . . . . . .  .  . . . .  . . 1 . . .  .  . .  . . . . . .  . . .
#> S1pr5     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Pdgfrb    . . . . . . . .  .  . . 1 .  . . . . . .  .  . .  1 . 5 . . .  . . .
#> Flt1      . . . . . . . .  .  . . . .  . . . . 1 .  .  . .  2 . 2 . . .  . . .
#> Chrm3     . . . . . . . .  .  . . . .  1 . . . . .  1  1 .  . . . . . .  2 . .
#> Agtr1a    . . . . . . . .  .  . . . .  . . . . . 1  .  . .  . . . . . .  . . .
#> Adgrl4    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ephb3     . . . . 1 . . .  .  . . 1 .  . . . . . .  .  . .  . . . . . .  . . .
#> Lpar6     . . . . . 1 . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Epha6     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Fzd5      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Emcn      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Cldn5     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Sstr1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Fzd3      . . . . . . . .  .  . 1 . .  . . . . . 1  1  . .  . . . . 1 .  . . .
#> P2ry12    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adgrl3    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Sema4d    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adgrf5    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr160    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Mrgpre    . . . . . . . .  .  . . . .  . 1 . . . .  .  1 .  . . . . . .  . . .
#> Amigo2    . . 1 . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adgra3    . . . . . . . .  .  . . . .  . . . . . 1  .  1 .  . . . . . .  . . .
#> P2yr13    . . . . 1 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Epha3     . . . . 1 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adgre1    . . . . 1 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> P2ry6     . . . . 1 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Cx3cl1    . . . . 1 . . .  .  . 1 . .  . . . . . .  .  . .  1 . . . . .  . . .
#> Epha10    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Tacr1     . . 1 . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Peg10     . . . . . . . .  1  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ednra     . . . . . . . .  .  . 1 . .  . . . . . .  .  . .  1 . 1 . . .  . . .
#> Plxnb3    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Flt4      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Slc17a7   . . . . . . . .  .  . . . .  . . . . . 1  .  . .  . . 1 . . 1  . . .
#> Drd2      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Tjap1     . . . . . . . .  .  . 1 . .  . . . . . .  .  . .  . . . . . .  . . .
#> Fzd8      . . . . . . . .  .  . . . .  . . . . . .  .  . .  1 . . . . .  . . .
#> Cd300c2   . . . . 1 . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Htr4      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr52     . . . . . . . .  1  . . . .  . . . . . .  .  . .  . . . . . .  1 . .
#> Tek       . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> S1pr2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Kdr       . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ror2      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Kcnj8     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . 2 . . .  . . .
#> Hrh1      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Celsr1    . . . . . . . .  .  . . . .  . . . . . .  .  . .  1 . . . . .  . . .
#> Gpr153    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . 1 . . .  . . .
#> S1pr3     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . 2 . . .  . . .
#> Adgra2    . . . . . . . .  .  . . . .  1 . . 2 . .  1  . .  . . 1 . . .  . . .
#> Fgfrl1    . . . . . . . .  .  . 2 1 .  . . . . . .  .  . .  . . . . . .  . . .
#> Tie1      . . . . . . . .  .  . . . .  . . . . . .  .  1 .  . . . . . .  . . .
#> Slc15a3   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ccr10     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Pdgfra    . . . . . . . .  .  . . . .  . . . . . .  .  . .  1 . . . . .  . . .
#> Lhcgr     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Mrgprf    . . . . . . . .  .  . . . .  . . . 1 . .  .  . .  . . . . . .  . . .
#> Ascl1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Baiap2    . . . . . . . .  .  . . . .  . . . . . .  1  . .  . . . . . .  . . .
#> Gpr85     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr22     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . 1 . . .  . . .
#> Gcgr      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Grm1      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ptafr     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Calcrl    . . . . . . . .  .  . 1 . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr34     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Cenpe     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adora2b   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adrb2     . . . . . . . .  .  . . . .  . . . . . .  .  2 .  . . . . . .  . . .
#> Ccr2      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  1 . .
#> Alk       . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Tacr3     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Chrm5     . . . . . . . .  .  . . . .  1 . . . . .  .  . .  . . . . . .  . . .
#> Erbb3     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ccr9      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr21     . . . . . . . .  .  . . . .  . 1 . . . .  .  . .  . . . . . .  . . .
#> Epha7     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adra2b    . . . . . . . .  .  . . . .  . . . . . .  .  . .  1 . . . . .  . . .
#> Pcdh15    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-104 . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adora2a   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Drd1      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-57  . . . . . . . .  .  . . . .  . . 1 . . .  .  . .  . . . . . .  . . .
#> Grm7      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Htr1b     . . . . . . . .  .  . . . .  . . 1 . . .  .  . .  . . . . . .  . . .
#> Fzd9      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Grm4      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-152 . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Aplnr     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr176    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> F2r       . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Vmn1r43   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr156    . . . . . . . .  .  . . . .  . . . 1 . .  .  . .  . . . . . .  . . .
#> Gpr55     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-155 . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Bdkrb2    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . 1 . . .  . . .
#> C5ar2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpbar1    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  1 . .
#> Adgrv1    . . . . . . . .  .  . . . .  1 . . . . .  .  3 .  . . . . . .  1 . .
#> Gprc5c    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Lmod1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr135    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  1 . .
#> Rxfp3     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Tbxa2r    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . 2 . . .  . . .
#> Fzd10     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ret       . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-80  . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Epha8     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr75     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . 1 . . .  . . .
#> Ackr1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adgrg4    . . . . . . . .  .  . . . .  . . . . . .  .  1 .  . . . . . .  . . .
#> Chrm1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . 1  . . .
#> Epha1     . . . . . . . .  .  . . . .  . . . . . .  .  1 .  . . . . . .  . . .
#> Vmn2r1    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Syt4      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Tas2r135  . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ccr1      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Lpar2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Myh11     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gipr      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Musk      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr183    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Cnr1      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> C3ar1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> S1pr4     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Grm2      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> C5ar1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ephb2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Fpr1      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Rho       . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr27     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adra1d    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr157    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adgrg2    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-45  . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr20     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Casr      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Avpr2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Mc5r      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Sctr      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Epha5     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr158    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Cxcr4     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Vipr1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr6      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-148 . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Celsr3    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Uts2r     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Hrh2      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Htr6      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Slc17a6   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr1      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Glp2r     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ppp1r3g   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> V1ra8     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . 1 . . . .  . . .
#> Trhr      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Sstr3     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr63     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Htr1a     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> F2rl3     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Pln       . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Grin2b    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr37     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Chrm4     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Lgr6      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr61     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Taar7e    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Mas1      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Npy2r     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ptger2    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> P2ry2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Oprm1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Prokr2    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adra1a    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr45     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Mc4r      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-128 . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Sstr2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Flt3      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Htr5a     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ghsr      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Galr1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-95  . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr83     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Crhr1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Brs3      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Drd5      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr139    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Taar6     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Htr5b     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Vmn1r40   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Rrh       . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-83  . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Crhr2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-161 . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Vmn1r50   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Tas1r3    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Taar2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> P2ry4     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Mchr1     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Tas1r1    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-69  . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Htr7      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Hcrtr2    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Htr2a     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Cysltr2   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Vmn1r46   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Grpr      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adgrf2    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-127 . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Opn5      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Taar3     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Oxtr      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Cckbr     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Nmbr      . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Ptger4    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Blank-145 . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Gpr101    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Slc17a8   . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Adgrf4    . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#> Epha2     . . . . . . . .  .  . . . .  . . . . . .  .  . .  . . . . . .  . . .
#>                      
#> Mlc1      .  . ......
#> Gprc5b    .  . ......
#> Gfap      .  1 ......
#> Ednrb     .  . ......
#> Sox9      .  . ......
#> Aqp4      .  . ......
#> Gjc3      2  . ......
#> Sox8      2  . ......
#> Ntsr2     .  . ......
#> Adcyap1r1 .  . ......
#> Smo       .  1 ......
#> Gpr161    .  . ......
#> C1qb      .  . ......
#> Olig1     3  . ......
#> C1qa      .  . ......
#> Lmtk2     1  . ......
#> Pth1r     .  . ......
#> Ephb6     .  . ......
#> Gpr26     .  . ......
#> Gabbr1    .  1 ......
#> Adra2a    .  . ......
#> Gabbr2    .  . ......
#> Timp4     .  . ......
#> Abcc9     .  . ......
#> Csf1r     .  . ......
#> Ntrk2     .  . ......
#> Adgrg1    .  . ......
#> Ntrk3     .  . ......
#> Ddr1      .  . ......
#> Erbb4     .  . ......
#> Selplg    .  . ......
#> Cspg5     1  1 ......
#> Dlk1      .  . ......
#> Gpr37l1   1  . ......
#> Fzd1      .  . ......
#> Fgfr3     .  . ......
#> Slc25a18  .  . ......
#> Igf1r     .  3 ......
#> Adgrl1    .  . ......
#> S1pr1     .  . ......
#> Ryk       .  . ......
#> Grm5      1  . ......
#> Eomes     .  . ......
#> Axl       .  . ......
#> Efemp1    .  . ......
#> Adgrb3    .  . ......
#> Sstr4     .  . ......
#> Fn1       .  5 ......
#> Adora1    1  . ......
#> Erbb2     .  . ......
#> Gper1     .  . ......
#> Gramd3    .  . ......
#> Mertk     .  . ......
#> Slc32a1   .  . ......
#> Adgra1    .  . ......
#> Adgrl2    .  . ......
#> Ptgfr     .  . ......
#> Fzd2      .  . ......
#> Aldh1l1   .  . ......
#> Adgrg6    .  . ......
#> Egfr      .  . ......
#> Gpr162    .  . ......
#> Lpar4     .  . ......
#> Kit       .  . ......
#> Ptgdr     .  . ......
#> Gad1      .  . ......
#> Lgr4      .  . ......
#> Tyro3     1  . ......
#> Atp13a5   .  . ......
#> Ddr2      .  . ......
#> Rspo3     .  . ......
#> Arhgap29  .  2 ......
#> Ackr3     .  . ......
#> Gpr146    .  . ......
#> Fzd7      .  . ......
#> Slc47a1   .  . ......
#> Insr      .  1 ......
#> Ptger3    .  . ......
#> Bdkrb1    .  . ......
#> Traf4     .  . ......
#> Celsr2    .  . ......
#> Fzd4      .  . ......
#> Fgfr2     .  . ......
#> Cx3cr1    .  . ......
#> Anxa11    .  2 ......
#> Adgrb1    2  . ......
#> Lpar1     .  . ......
#> Gpr17     2  . ......
#> Gpr173    .  . ......
#> Ror1      .  . ......
#> Cxcl12    .  1 ......
#> Kiss1r    .  . ......
#> Epha4     .  . ......
#> Gpr182    .  . ......
#> Man1a     .  . ......
#> Gpr4      .  . ......
#> P2ry14    .  . ......
#> Hcar1     .  . ......
#> Ephb1     .  . ......
#> Adrb1     .  . ......
#> Grm3      .  . ......
#> P2ry1     .  . ......
#> Nrp2      .  . ......
#> Ptk7      .  . ......
#> Slco1a4   .  3 ......
#> Tmem108   .  . ......
#> Hrh3      .  . ......
#> Insrr     .  . ......
#> Cmklr1    .  . ......
#> Opn3      .  . ......
#> Adra1b    .  . ......
#> Gpr62     .  . ......
#> Npy1r     .  . ......
#> Nrp1      .  . ......
#> Fzd6      .  2 ......
#> S1pr5     .  . ......
#> Pdgfrb    .  . ......
#> Flt1      .  8 ......
#> Chrm3     .  . ......
#> Agtr1a    .  . ......
#> Adgrl4    .  . ......
#> Ephb3     .  . ......
#> Lpar6     .  . ......
#> Epha6     .  . ......
#> Fzd5      .  . ......
#> Emcn      .  . ......
#> Cldn5     . 10 ......
#> Sstr1     1  . ......
#> Fzd3      .  . ......
#> P2ry12    .  . ......
#> Adgrl3    .  . ......
#> Sema4d    .  . ......
#> Adgrf5    .  2 ......
#> Gpr160    .  . ......
#> Mrgpre    .  . ......
#> Amigo2    .  . ......
#> Adgra3    .  . ......
#> P2yr13    .  . ......
#> Epha3     .  . ......
#> Adgre1    .  . ......
#> P2ry6     .  . ......
#> Cx3cl1    .  . ......
#> Epha10    .  . ......
#> Tacr1     .  . ......
#> Peg10     .  . ......
#> Ednra     .  . ......
#> Plxnb3    .  . ......
#> Flt4      .  1 ......
#> Slc17a7   .  . ......
#> Drd2      .  . ......
#> Tjap1     .  . ......
#> Fzd8      .  . ......
#> Cd300c2   .  . ......
#> Htr4      .  . ......
#> Gpr52     .  . ......
#> Tek       .  . ......
#> S1pr2     .  . ......
#> Kdr       .  . ......
#> Ror2      .  . ......
#> Kcnj8     .  1 ......
#> Hrh1      .  . ......
#> Celsr1    .  . ......
#> Gpr153    .  . ......
#> S1pr3     .  . ......
#> Adgra2    .  1 ......
#> Fgfrl1    1  . ......
#> Tie1      .  1 ......
#> Slc15a3   .  . ......
#> Ccr10     .  . ......
#> Pdgfra    .  . ......
#> Lhcgr     .  . ......
#> Mrgprf    .  . ......
#> Ascl1     .  . ......
#> Baiap2    1  . ......
#> Gpr85     .  . ......
#> Gpr22     .  . ......
#> Gcgr      .  . ......
#> Grm1      .  . ......
#> Ptafr     .  . ......
#> Calcrl    .  . ......
#> Gpr34     .  . ......
#> Cenpe     .  . ......
#> Adora2b   .  . ......
#> Adrb2     .  . ......
#> Ccr2      .  . ......
#> Alk       .  . ......
#> Tacr3     .  . ......
#> Chrm5     .  . ......
#> Erbb3     .  . ......
#> Ccr9      .  . ......
#> Gpr21     .  . ......
#> Epha7     .  . ......
#> Adra2b    .  . ......
#> Pcdh15    1  . ......
#> Blank-104 .  . ......
#> Adora2a   .  . ......
#> Drd1      .  . ......
#> Blank-57  .  . ......
#> Grm7      .  . ......
#> Htr1b     .  . ......
#> Fzd9      .  . ......
#> Grm4      .  . ......
#> Blank-152 .  . ......
#> Aplnr     .  . ......
#> Gpr176    .  . ......
#> F2r       .  . ......
#> Vmn1r43   .  . ......
#> Gpr156    .  . ......
#> Gpr55     .  . ......
#> Blank-155 .  . ......
#> Bdkrb2    .  . ......
#> C5ar2     .  . ......
#> Gpbar1    .  . ......
#> Adgrv1    .  . ......
#> Gprc5c    1  . ......
#> Lmod1     .  . ......
#> Gpr135    .  . ......
#> Rxfp3     .  . ......
#> Tbxa2r    .  . ......
#> Fzd10     .  . ......
#> Ret       .  . ......
#> Blank-80  .  . ......
#> Epha8     .  . ......
#> Gpr75     .  . ......
#> Ackr1     .  . ......
#> Adgrg4    .  . ......
#> Chrm1     .  . ......
#> Epha1     .  . ......
#> Vmn2r1    .  . ......
#> Syt4      .  . ......
#> Tas2r135  .  . ......
#> Ccr1      .  . ......
#> Lpar2     .  . ......
#> Myh11     .  . ......
#> Gipr      .  . ......
#> Musk      .  . ......
#> Gpr183    .  . ......
#> Cnr1      .  . ......
#> C3ar1     .  . ......
#> S1pr4     .  . ......
#> Grm2      .  . ......
#> C5ar1     .  . ......
#> Ephb2     .  . ......
#> Fpr1      .  . ......
#> Rho       .  . ......
#> Gpr27     .  . ......
#> Adra1d    .  . ......
#> Gpr157    .  . ......
#> Adgrg2    .  . ......
#> Blank-45  .  . ......
#> Gpr20     .  . ......
#> Casr      .  . ......
#> Avpr2     .  . ......
#> Mc5r      .  . ......
#> Sctr      .  . ......
#> Epha5     .  . ......
#> Gpr158    .  . ......
#> Cxcr4     .  . ......
#> Vipr1     .  . ......
#> Gpr6      .  . ......
#> Blank-148 .  . ......
#> Celsr3    .  . ......
#> Uts2r     .  . ......
#> Hrh2      .  . ......
#> Htr6      .  . ......
#> Slc17a6   .  . ......
#> Gpr1      .  . ......
#> Glp2r     .  . ......
#> Ppp1r3g   .  . ......
#> V1ra8     .  . ......
#> Trhr      .  . ......
#> Sstr3     .  . ......
#> Gpr63     .  . ......
#> Htr1a     .  . ......
#> F2rl3     .  . ......
#> Pln       .  . ......
#> Grin2b    .  . ......
#> Gpr37     .  . ......
#> Chrm4     .  . ......
#> Lgr6      .  . ......
#> Gpr61     .  . ......
#> Taar7e    .  . ......
#> Mas1      .  . ......
#> Npy2r     .  . ......
#> Ptger2    .  . ......
#> P2ry2     .  . ......
#> Oprm1     .  . ......
#> Prokr2    .  . ......
#> Adra1a    .  . ......
#> Gpr45     .  . ......
#> Mc4r      .  . ......
#> Blank-128 .  . ......
#> Sstr2     .  . ......
#> Flt3      .  . ......
#> Htr5a     .  . ......
#> Ghsr      .  . ......
#> Galr1     .  . ......
#> Blank-95  .  . ......
#> Gpr83     .  . ......
#> Crhr1     .  . ......
#> Brs3      .  . ......
#> Drd5      .  . ......
#> Gpr139    .  . ......
#> Taar6     .  . ......
#> Htr5b     .  . ......
#> Vmn1r40   .  . ......
#> Rrh       .  . ......
#> Blank-83  .  . ......
#> Crhr2     .  . ......
#> Blank-161 .  . ......
#> Vmn1r50   .  . ......
#> Tas1r3    .  . ......
#> Taar2     .  . ......
#> P2ry4     .  . ......
#> Mchr1     .  . ......
#> Tas1r1    .  . ......
#> Blank-69  .  . ......
#> Htr7      .  . ......
#> Hcrtr2    .  . ......
#> Htr2a     .  . ......
#> Cysltr2   .  . ......
#> Vmn1r46   .  . ......
#> Grpr      .  . ......
#> Adgrf2    .  . ......
#> Blank-127 .  . ......
#> Opn5      .  . ......
#> Taar3     .  . ......
#> Oxtr      .  . ......
#> Cckbr     .  . ......
#> Nmbr      .  . ......
#> Ptger4    .  . ......
#> Blank-145 .  . ......
#> Gpr101    .  . ......
#> Slc17a8   .  . ......
#> Adgrf4    .  . ......
#> Epha2     .  . ......
#> 
#>  .....suppressing 465 columns in show(); maybe adjust options(max.print=, width=)
#>  ..............................

dim(ovlp)
#> [1] 15050     3
nrow(ovlp) # number of relationships
#> [1] 15050

# get feature unique IDs overlapped by nth poly
ovlp[1] # check one (no overlaps returns integer(0))
#> integer(0)
ovlp[1:5] # check multiple
#>  [1]   87  122  138  151  154  177  220  228  231  436  459  529  669  670  848
#> [16]  941  954  967  973  975  976  981  987  996  998  999 1010 1016 1019 1021
#> [31] 1024 1034 1036 1037 1040 1043 1044 1047 1051 1053 1055 1056 1059 1064 1067
#> [46] 1070 1072 1076 1080 1086 1089 1098 1100 1104 1106 1108 1111 1115 1116 1121
#> [61] 1122 1129 1132 1137 1143 1151 1152 1154 1155 1157 1163 1173 1176 1180 1181
#> [76] 1183 1185 1188 1199 1203 1205 1225 1229 1244 1251 1254 1260 1267 1274 1282
ovlp[1:5, use_names = TRUE] # returns feature names, but no longer unique
#>  [1] "Selplg"    "Fgfr3"     "Gfap"      "Aqp4"      "Adcyap1r1" "Ntrk2"    
#>  [7] "C1qa"      "Egfr"      "Olig1"     "Ntrk2"     "Sox9"      "Ntsr2"    
#> [13] "Adgrb1"    "Gfap"      "Slc25a18"  "Adgrg1"    "Gprc5b"    "Ntsr2"    
#> [19] "Smo"       "Slc25a18"  "Gprc5b"    "Nrp1"      "Axl"       "Slc32a1"  
#> [25] "Amigo2"    "C1qa"      "Ntsr2"     "C1qa"      "C1qb"      "C1qb"     
#> [31] "Gpr37l1"   "P2yr13"    "Mertk"     "Mertk"     "Epha3"     "Adgre1"   
#> [37] "Ntsr2"     "Timp4"     "Cx3cr1"    "P2ry6"     "C1qa"      "C1qb"     
#> [43] "C1qa"      "C1qa"      "Ntsr2"     "Ntsr2"     "Cx3cl1"    "Aldh1l1"  
#> [49] "Ntsr2"     "Adgrg1"    "Ntrk2"     "Gpr37l1"   "Ntsr2"     "Adrb1"    
#> [55] "Tacr1"     "Selplg"    "C1qa"      "Fgfr3"     "Ntrk2"     "C1qa"     
#> [61] "Ednrb"     "Celsr2"    "Ntsr2"     "Sox8"      "Mertk"     "Ntsr2"    
#> [67] "Fgfr3"     "C1qb"      "Ryk"       "Cx3cr1"    "Gpr146"    "Gpr37l1"  
#> [73] "Ephb3"     "C1qb"      "Cx3cr1"    "C1qb"      "Ednrb"     "C1qa"     
#> [79] "Celsr2"    "Aqp4"      "Cx3cr1"    "Mertk"     "Csf1r"     "Sox8"     
#> [85] "Selplg"    "C1qa"      "Ednrb"     "Timp4"     "Sox8"      "Cd300c2"  

# get integer index of poly(s) overlapping particular feature species
ovlp[, 1]
#>  [1]  49 126 115  73 140 133  83 136 150  37  91 124  44 178 354 203 393 314 330
#> [20] 313 319 243 276 288 312 289 211 384 397 157 237 324 332 339 285 327 270 497
#> [39] 454 486 498 470 474 446 491 475 485
ovlp[, "Mlc1"] # this is the same
#>  [1]  49 126 115  73 140 133  83 136 150  37  91 124  44 178 354 203 393 314 330
#> [20] 313 319 243 276 288 312 289 211 384 397 157 237 324 332 339 285 327 270 497
#> [39] 454 486 498 470 474 446 491 475 485

# get a subset of overlap object
ovlp[1:10, ids = FALSE] # subset to first 10 polys
#> <overlapPointDT>
#> spat_unit : "z0"
#> feat_type : "rna"
#> provenance: z0 
#> * polygons : 10
#> * features : 35889
#> * relations: 202
ovlp[, 1:10, ids = FALSE] # subset to first 10 feature species
#> <overlapPointDT>
#> spat_unit : "z0"
#> feat_type : "rna"
#> provenance: z0 
#> * polygons : 498
#> * features : 35889
#> * relations: 1606
ovlp[1:10, 1:10] # subset to first 10 polys and first 10 features species
#> <overlapPointDT>
#> spat_unit : "z0"
#> feat_type : "rna"
#> provenance: z0 
#> * polygons : 10
#> * features : 35889
#> * relations: 65