overlapToMatrix

create a count matrix based on overlap results from calculateOverlap

Usage

# S4 method for class 'giotto'
overlapToMatrix(
  x,
  name = "raw",
  spat_info = NULL,
  feat_info = NULL,
  type = c("point", "intensity"),
  feat_count_column = NULL,
  fun = "sum",
  return_gobject = TRUE,
  verbose = TRUE,
  aggr_function = deprecated(),
  poly_info = deprecated(),
  count_info_column = deprecated(),
  ...
)

# S4 method for class 'giottoPolygon'
overlapToMatrix(
  x,
  feat_info = "rna",
  type = c("point", "intensity"),
  feat_count_column = NULL,
  count_info_column = deprecated(),
  ...
)

# S4 method for class 'SpatVector'
overlapToMatrix(
  x,
  col_names = NULL,
  row_names = NULL,
  feat_count_column = NULL,
  output = c("Matrix", "data.table"),
  verbose = TRUE,
  count_info_column = deprecated(),
  ...
)

# S4 method for class 'data.table'
overlapToMatrix(
  x,
  fun = "sum",
  output = c("Matrix", "data.table"),
  aggr_function = deprecated()
)

# S4 method for class 'overlapPointDT'
overlapToMatrix(
  x,
  name = "raw",
  sort = TRUE,
  feat_count_column = NULL,
  output = c("Matrix", "exprObj"),
  ...
)

# S4 method for class 'overlapIntensityDT'
overlapToMatrix(
  x,
  name = "raw",
  sort = TRUE,
  output = c("Matrix", "exprObj"),
  ...
)

Arguments

x

object containing overlaps info. Can be giotto object or SpatVector points or data.table of overlaps generated from calculateOverlap

name

name for the overlap count matrix

spat_info

character. Polygon information to use

feat_info

character. Feature information to use

type

character. Type of overlap data (either 'point' or 'intensity')

feat_count_column

column with count information. If a column called "count" is present in the feature points data, it will be automatically selected.

fun

character. Function to aggregate image information (default = "sum")

return_gobject

return giotto object (default: TRUE)

verbose

be verbose

aggr_function

deprecated. Use fun instead.

poly_info

deprecated. Please use spat_info.

count_info_column

deprecated. Use feat_count_column instead.

...

additional params to pass to methods

col_names, row_names

character vector. (optional) Set of row and col names that are expected to exist. This fixes the dimensions of the matrix since the overlaps information does not directly report rows and cols where no values were detected.

output

data format/class to return the results as. Default is "Matrix"

sort

logical (default = TRUE). Whether to perform a mixed sort on output matrix row and col names.

Value

giotto object or count matrix

Examples

g <- GiottoData::loadGiottoMini("vizgen")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 4. read Giotto image information
#> python already initialized in this session
#>  active environment : '/usr/bin/python3'
#>  python version : 3.12
gpoly <- getPolygonInfo(g,
    polygon_name = "aggregate",
    return_giottoPolygon = TRUE
)
gpoints <- getFeatureInfo(g, return_giottoPoints = TRUE)

# calculate all transcripts overlapped
out_all <- calculateOverlap(gpoly, gpoints)

overlapToMatrix(out_all)
#> 337 x 462 sparse Matrix of class "dgCMatrix"
#>   [[ suppressing 33 column names ‘1132915601442719251817312578799507532’, ‘1900302715660571356090444774019116326’, ‘2467888014556719520437642348850497467’ ... ]]
#>                                                                                
#> Abcc9     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . 5 .  . . .  .
#> Ackr1     .  . . 1 . . 1 . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ackr3     .  . . . . . . . .  .  .  .  2 . . .  .  . . . .  . 2 . . .  . . .  2
#> Adcyap1r1 3  . 2 2 . . . . 3  1  4  1  5 . 2 3  2  2 . 1 .  2 . . 1 .  3 . .  .
#> Adgra1    .  . . 1 . . . . 1  .  1  .  . . . .  .  2 . . .  . . . . .  . 1 .  .
#> Adgra2    .  2 . . . . 1 . .  .  .  .  . . . .  .  . 1 . .  . . . 1 .  . . .  .
#> Adgra3    .  . 1 . . . 1 . .  1  1  .  . . . .  .  . . . .  . . . . .  1 . .  1
#> Adgrb1    1  . 7 1 1 . . 1 6  1  4  .  1 . . 1  3  . . . .  4 . . . .  . . .  .
#> Adgrb3    2  . 1 . . . . . 1  3  1  2  . . 5 .  .  1 . . .  3 . . . .  . . .  .
#> Adgre1    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Adgrf2    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Adgrf4    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Adgrf5    .  . . . . . 1 . .  .  .  .  . 1 . .  .  1 . . .  . . . . .  . . .  .
#> Adgrg1    .  . . . . . 2 2 .  .  .  3  . 2 . .  .  . . . .  . . . . .  1 1 .  .
#> Adgrg2    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  1
#> Adgrg4    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Adgrg6    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Adgrl1    1  . . . 2 1 . . .  2  .  .  . . 1 3  1  4 . . 1  7 . . . 1  . 1 .  1
#> Adgrl2    .  . . . . 2 . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Adgrl3    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . 1 .  . . .  .
#> Adgrl4    .  . . . . . 1 . .  .  .  .  . 1 . .  .  . . . .  . . . 1 .  . . .  .
#> Adgrv1    .  . . . . . . . .  .  .  .  . . . .  .  . 1 . .  . . . . .  . . .  .
#> Adora1    .  . 1 . . . . . .  2  .  .  . . . .  2  1 . . .  . . . . .  . . .  .
#> Adora2a   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Adora2b   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Adra1a    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  1 . . . .  . . .  .
#> Adra1b    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Adra1d    1  . . . . . . . 2  1  2  .  . . . .  2  1 . . .  1 . . . .  . . .  .
#> Adra2a    1  . 1 . . . . . .  .  .  1  1 . . .  .  . . . .  . . . . .  1 . .  1
#> Adra2b    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . 1 . . .  . . .  .
#> Adrb1     .  . . . . . . 1 .  .  .  .  . . . .  .  . . . .  1 . . . .  1 . .  .
#> Adrb2     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Agtr1a    .  . 1 . . . . . .  .  .  .  . . . .  .  1 . . .  . . . . .  2 . .  .
#> Aldh1l1   .  . . . . 1 . . .  2  .  4  . . 1 .  .  . 1 . .  . . . . .  1 . 1  1
#> Alk       .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  4
#> Amigo2    1  . . . . . . . 2  .  .  .  . . . .  .  . . . .  1 . . . .  . . .  .
#> Anxa11    .  . . . . . . . .  .  1  .  . . . 1  .  . . . .  1 1 1 1 .  . . .  1
#> Aplnr     .  . . . . . . . .  .  1  .  2 . . .  .  . . . .  . . . . .  . . .  .
#> Aqp4      .  . . . . . . 1 .  1  .  9 19 . 2 .  1  . . . .  . . . . 1 11 1 1  .
#> Arhgap29  .  . . . . 1 . . .  1  .  .  . . . .  .  1 1 . .  . 3 3 2 .  . . 2  6
#> Ascl1     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Atp13a5   .  . . . . . . . .  .  .  1  . . . .  .  . 3 . .  . 5 . 9 .  . . 4 10
#> Avpr2     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  1
#> Axl       .  . . . . 1 . . .  1  .  .  1 . . .  .  . . . .  . . . . .  3 . .  1
#> Baiap2    5  . 3 . . 1 . 1 4  4  5  .  . . . 1  2  3 2 3 .  6 . 1 . .  . . .  .
#> Bdkrb1    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Bdkrb2    .  . . . . 1 . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-161 .  . . . 1 . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  1
#> Blank-155 .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-152 .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-148 .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-145 .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-139 .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-128 .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-127 .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-104 .  . . . . . . . .  .  1  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-95  .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-83  .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-80  .  . . . . . . . .  .  1  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-69  .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Blank-57  .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . 1 . . .  . . .  .
#> Blank-45  .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Brs3      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> C1qa      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> C1qb      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> C3ar1     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> C5ar1     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> C5ar2     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Calcrl    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  2 . . . .  . . .  .
#> Casr      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Cckbr     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ccr1      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ccr2      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ccr9      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ccr10     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Cd300c2   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Celsr1    .  . . . . . . . .  .  .  2  . . . .  .  . . . .  . . . . .  . . .  .
#> Celsr2    2  . 2 . . . . . .  1  4  .  . . 1 1  4  . . . .  3 . . . .  1 . .  .
#> Celsr3    .  . . . . . . . 1  .  .  .  . . . .  .  1 . . .  . . . . .  . . .  .
#> Cenpe     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Chrm1     3  . 2 1 3 . . . 1  .  1  .  . . . 3  1  1 . 1 2  6 . . . .  . . .  .
#> Chrm3     .  . . . . . . . .  .  4  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Chrm4     .  . . 1 . . . . 1  .  3  .  . . . .  .  . . 1 .  . . . . .  . . .  .
#> Chrm5     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Cldn5     .  . . . . . 9 . .  .  .  .  . 3 . .  .  . . . 1  . . . 1 .  . . .  .
#> Cmklr1    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Cnr1      .  . . . . . . . .  . 10  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Crhr1     .  . . . . . . . 1  .  1  .  . . 1 .  .  . . . .  . . . . .  . . .  .
#> Crhr2     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Csf1r     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Cspg5     .  . . . 1 . . 6 .  4  2  6  . . 5 .  .  1 . . .  1 . . 1 1  4 . 1  .
#> Cx3cl1    6  . 4 3 . 1 . 2 4  4  3  .  1 . . 1  4  6 . 1 2  8 . . . .  . . 2  3
#> Cx3cr1    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Cxcl12    .  2 . . . 2 1 . .  .  .  3  . . . 1  3  . 2 . .  1 2 1 . .  . . 5  5
#> Cxcr4     .  . . . . . . . .  .  .  .  1 . . .  .  . . . .  . . . . .  . . .  .
#> Cysltr2   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . 2 .  . . .  .
#> Ddr1      .  . . . . . . . .  1  1  2  1 . 2 .  .  2 . . .  . . . . .  . . .  .
#> Ddr2      .  1 . . . . 1 . .  .  .  .  . . . .  .  . . . .  . . 1 . .  . . .  2
#> Dlk1      .  . . . . . . . .  .  .  .  1 . . .  .  . . . .  . . . . .  . . .  .
#> Drd1      .  . . 1 . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Drd2      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Drd5      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ednra     1  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ednrb     .  . . . . . . . .  1  .  2  2 . 4 .  .  . . . .  . . . . .  1 . 1  .
#> Efemp1    .  . . . . . . . .  .  .  3  . . . .  .  . 4 . .  . 1 . 1 .  1 . 1  1
#> Egfr      .  . 1 . . . . . .  .  .  .  . . . .  .  1 . . .  . . . . .  . . .  .
#> Emcn      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Eomes     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Epha1     .  . . 1 1 . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Epha2     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Epha3     1  . . . . . . . .  .  1  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Epha4     3  1 2 2 1 . . . 3  3  1  1  . 1 . .  5  2 . 1 .  1 . . . .  . . 1  .
#> Epha5     1  . . . . . 1 . .  1  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Epha6     2  . . 6 1 . . . 3  1  6  .  . . . .  1  . . . .  2 . . 1 .  . . .  .
#> Epha7     2  . . . 1 . . . .  3  1  .  . . . .  .  1 . . .  1 . . . .  . . .  .
#> Epha8     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Epha10    .  . 1 . 1 . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ephb1     .  . . . . . . . .  .  .  .  . . 1 .  .  1 . . .  . . . . .  . 1 .  .
#> Ephb2     .  . . 1 . . . . .  1  .  .  . . . .  .  . . . .  . . . . .  . . .  1
#> Ephb3     1  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  2
#> Ephb6     .  . 1 . . . . . 1  .  1  1  . . . .  1  . . 2 1  2 1 . . .  1 . .  .
#> Erbb2     .  1 . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Erbb3     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Erbb4     1  . . . . . . . 1  .  .  .  . . . .  .  . . . .  . . . . .  . 1 .  .
#> F2r       .  . . . . . . . 1  .  .  .  . . . .  .  . . . .  . . . 1 .  . . .  .
#> F2rl3     .  1 . . . . . . .  .  1  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Fgfr2     .  1 . . . 1 . . .  .  .  .  1 . 1 .  .  . . . .  . 2 . 1 .  1 . 1  .
#> Fgfr3     .  . . 1 . . . 2 .  1  .  6  2 . 3 .  .  . . . .  . . . 1 .  6 1 .  .
#> Fgfrl1    .  . . . . . . 2 .  .  .  .  . . . .  .  . 1 . .  . . . . .  . . .  .
#> Flt1      .  . . . . . 2 . .  .  .  .  . 1 . .  .  1 . . .  . . . 5 .  . . .  .
#> Flt3      1  . 1 . . . . . .  .  1  .  . . . .  .  1 . . .  . . . . .  . . .  .
#> Flt4      .  . . . . . 1 . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Fn1       .  2 . . . . . . .  .  .  .  . 2 . .  .  . 9 . .  . 6 . 2 .  . . 3  2
#> Fpr1      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Fzd1      .  1 . . . . . . .  .  .  .  1 . . .  .  . 1 . .  . . . . .  . . .  .
#> Fzd2      .  1 . . . 2 . . .  .  .  .  . . . .  .  . 4 . .  . 1 . . .  . . .  2
#> Fzd3      4  . 1 . . . . . .  1  .  .  1 . . 1  .  . . . .  3 . . . .  2 . .  .
#> Fzd4      .  . . . . . . . .  .  .  .  . . . .  .  1 . . .  . . . . .  . . .  .
#> Fzd5      .  . . . . . . . .  .  .  .  1 . . .  .  . . . .  . . . . .  . . .  .
#> Fzd6      .  . . . . . . . .  .  .  .  . . . .  .  . 1 . .  . 1 . . .  . . .  1
#> Fzd7      2  . . . . . . . .  .  .  .  . . . .  .  . 2 . .  . 1 1 . .  . . 2  1
#> Fzd8      .  1 . . 1 . . . .  .  .  .  . . . .  .  . . . .  . . . 1 .  1 . .  1
#> Fzd9      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Fzd10     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gabbr1    6  1 2 2 3 . 2 2 3  .  4  1  3 . 1 4  2  2 . 3 .  6 . . . .  3 1 .  .
#> Gabbr2    1  . 2 1 . . 1 . 1  1  1  .  . . . 2  2  2 . . .  6 . . . .  . . .  .
#> Gad1      2  . 2 . . . . 1 .  .  4  .  . . . .  .  1 . . .  4 . . . .  1 . .  .
#> Galr1     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gcgr      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gfap      .  . . . . . . 5 .  2  1  5 43 . 1 .  .  . . 1 3  . . . 3 4  8 4 .  .
#> Ghsr      1  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gipr      .  . . . . . . . .  .  1  .  . . . .  .  . . . .  . . . 1 .  . . .  .
#> Gjc3      .  . . . . . . . .  .  .  .  3 . . .  .  . . . .  . . . . .  . . .  .
#> Glp2r     2  . . . . . . . .  3  .  .  . . . 1  1  . . . .  . . . . .  . . .  .
#> Gpbar1    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gper1     .  1 . . . . . . .  .  .  .  . . . .  .  . . . .  . . . 1 .  . . .  .
#> Gpr1      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr4      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  1 . .  .
#> Gpr6      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr17     .  . . . . . . . .  .  .  .  . . . 1  .  . . . .  . . . . .  . 1 .  .
#> Gpr20     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr21     .  . 1 . . . . . .  .  2  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr22     1  . 1 2 . . . . 1  2  1  .  . . . .  1  . . 1 .  . . . . .  . . .  .
#> Gpr26     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr27     .  . . . . . . . .  1  .  .  . . . .  .  . . . 1  2 . . . .  . . .  .
#> Gpr34     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr37     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr37l1   .  . . . . . . 2 .  2  2 11  8 . 9 .  .  . . . .  . . . 2 . 14 1 .  .
#> Gpr45     .  . . . . . . . .  .  .  .  . . . .  .  . . 1 .  . . . . .  . . .  .
#> Gpr52     .  . . . . . . . .  .  .  1  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr55     .  . . . . . . . .  .  .  .  . . . .  .  . 1 . .  . . . . .  . . .  .
#> Gpr61     .  . . . . . . . .  1  .  .  . . . .  .  . . 2 1  1 . . . .  . . .  .
#> Gpr62     1  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr63     .  . 1 . . . . . 1  .  .  .  . . . .  1  . . . .  . . . . .  . . .  .
#> Gpr75     .  . . . . . . . .  2  1  .  . . . .  .  . . . .  . . . . .  . . .  1
#> Gpr83     .  . . 1 . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr85     .  . . . 1 . . . 2  .  .  .  . . . .  1  . . . .  2 . . . .  . . .  .
#> Gpr101    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr135    .  . . 1 . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr139    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr146    .  . . . 1 1 1 1 .  .  .  .  . . . .  .  . 1 . .  . . . . .  . . 2  2
#> Gpr153    .  . . . . . 1 . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr156    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr157    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gpr158    1  . . 1 . . . . 1  .  5  .  . . . .  .  . . 1 .  2 . . . .  . . .  .
#> Gpr160    .  . . . . . . . .  .  .  .  . . 1 .  .  . . . .  . . . . .  . . .  .
#> Gpr161    .  . . . . . . . .  .  .  .  . . 1 .  .  . . . .  1 . . . .  . . .  .
#> Gpr162    1  . 2 1 1 . . 1 1  .  5  .  1 . 3 .  3  . . 1 .  3 . . . .  1 . .  .
#> Gpr173    .  . . . . . . . 1  .  .  .  . . . .  .  . . . .  . . . . .  . . .  1
#> Gpr176    .  . . . . . . . .  .  1  .  . . . .  1  1 . . .  . . . . .  . . .  .
#> Gpr182    .  . . . . . . . .  .  .  1  . . . .  .  . 1 . .  . . . . .  . . .  5
#> Gpr183    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Gprc5b    .  . 1 . . . . 1 .  .  .  3  . . 1 .  1  1 . . .  1 1 . 1 1  . . .  .
#> Gprc5c    .  . . . . 2 . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  3
#> Gramd3    .  . . . . . . . .  .  .  .  . . . .  .  . 1 . .  . . . . .  . . .  .
#> Grin2b    2  . 2 2 . . . . 4  .  1  .  . . . 5  1  2 . 3 .  5 . . . .  . . .  .
#> Grm1      .  . 2 1 . . . . 3  2  1  .  . . . 1  1  1 . 1 .  . . . . .  . . .  .
#> Grm2      1  . . . . . . . 1  1  .  .  . . . .  .  . . . .  1 . . . .  . . .  .
#> Grm3      .  . . . . 1 . . .  .  .  .  . . 1 .  .  . . . .  . . . 3 .  1 . .  .
#> Grm4      .  . . . . . . . .  .  1  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Grm5      5  . 1 1 3 . . . 1  1  1  .  . . . 5  2  2 . 2 .  7 . . . .  1 . .  .
#> Grm7      1  . 2 . . . . . 4  4  7  .  . . . 3  2  3 . . .  2 . . 1 .  . . .  .
#> Grpr      .  1 . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Hcar1     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . 1 . . .  . . .  5
#> Hcrtr2    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Hrh1      .  . . . . . . 1 .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Hrh2      .  . . . . . . . .  .  1  .  . . . .  .  . . . .  1 . . . .  . . .  .
#> Hrh3      .  . . . . . . 1 .  1  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Htr1a     .  . 1 . . . . . .  .  .  .  . . . .  .  . . . .  1 . . . .  . . .  .
#> Htr1b     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . 1 . . .  . . .  .
#> Htr2a     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Htr4      .  . 2 . . . . . .  1  1  .  . . . .  1  1 . . .  1 . . . .  . . .  .
#> Htr5a     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Htr5b     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Htr6      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  1 . .  .
#> Htr7      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Igf1r     1  . . 1 2 . . 1 1  .  2  1  . 3 . 1  .  1 . . .  . . . 1 .  . . .  .
#> Insr      .  . 1 1 2 . 1 . 1  .  1  .  . . . 1  1  1 1 . .  2 . . . .  . . .  .
#> Insrr     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  5
#> Kcnj8     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . 8 .  . . .  .
#> Kdr       .  . . . . . . . .  .  .  .  . . . .  .  1 . . .  . . . 1 .  . . .  .
#> Kiss1r    .  . . 1 . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Kit       .  . . . . 3 . . .  .  .  .  . . . .  .  . 1 . .  . 1 . . .  . . .  8
#> Lgr4      .  . 1 . . 1 . . .  1  .  .  . . 1 1  .  . 2 . .  . 1 . . .  3 . .  1
#> Lgr6      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Lhcgr     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Lmod1     .  1 . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Lmtk2     2  . 1 . 1 . . . 1  1  1  .  . . . .  3  2 . . .  1 . . . .  . . .  .
#> Lpar1     .  2 . . . 1 . . .  .  .  2  . . . .  .  . 2 . .  . 1 . . .  . . 1  1
#> Lpar2     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Lpar4     .  . . . . 1 . . .  1  1  .  . . . .  .  1 . . .  . . . . .  . . .  .
#> Lpar6     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . 1  .
#> Man1a     .  . . . . 1 . . .  .  .  .  . . . .  .  . 2 . .  . . . . .  1 . .  .
#> Mas1      .  . 2 2 . . . . 1  .  3  .  . . . 1  .  . . 2 .  1 . . . .  . . .  .
#> Mc4r      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Mc5r      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Mchr1     .  . . . . . . . .  .  1  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Mertk     .  . . . . . . . .  .  .  1  . . 3 .  .  . . . .  . . . . .  3 . .  1
#> Mlc1      1  . . 1 . . . 3 .  2  .  1  2 . 3 .  .  . . . 1  3 . . . .  . . .  .
#> Mrgpre    .  . . . . 1 . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . 1  .
#> Mrgprf    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . 1 . . .  . . .  .
#> Musk      .  . . . . . . . .  .  1  1  . . . .  .  . . . .  . . . . .  . . .  .
#> Myh11     . 21 . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Nmbr      .  . . . . . . 1 .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Npy1r     .  . 1 . . 1 . . .  .  .  .  . . . .  .  . 1 . .  . 1 . . .  . . .  1
#> Npy2r     2  . 1 . 1 . . . .  2  .  .  . . . .  1  1 . . .  3 . . . .  . . .  .
#> Nrp1      1  . 1 1 1 . . . 1  .  1  1  . . . 2  .  3 . . .  1 . . . .  1 . .  .
#> Nrp2      .  . . . 1 . . 1 1  .  .  1  . . . .  .  . 1 . .  . . . . .  . . .  .
#> Ntrk2     .  . . . . . . 3 .  3  2  4  5 1 6 .  .  . . . 1  1 . . . .  4 3 .  .
#> Ntrk3     5  3 4 1 3 2 . . 6  2  4  .  1 . . 3  3  . . 3 1  5 . . . .  1 . .  3
#> Ntsr2     .  . . . . . . 4 .  .  1  6  7 1 4 .  .  . . . .  1 . 1 1 1 12 3 .  .
#> Olig1     .  . . . . . . . .  1  .  .  1 . . .  1  . . . .  . . . . 1  . 2 .  .
#> Opn3      .  . . . . . . . .  .  .  .  . . . .  .  . 1 . .  . . . . .  . . .  2
#> Opn5      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Oprm1     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Oxtr      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> P2ry1     .  . . . . . . . .  .  .  1  . . . .  .  . . . .  . 2 . . .  . . .  2
#> P2ry2     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> P2ry4     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> P2ry6     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> P2ry12    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> P2ry14    .  . . . . . . . 1  .  .  1  . . . .  .  . . . .  . . . 3 .  . . 2  .
#> P2yr13    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Pcdh15    .  . . . . . . . .  1  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Pdgfra    .  1 . . . 1 . . .  .  .  .  . . . .  .  . . . .  . 1 . . 1  . . .  1
#> Pdgfrb    .  . . . . 3 . . .  .  .  .  . . . .  .  . . . .  . 1 . 2 .  . . .  4
#> Peg10     .  . . . . . . . .  .  2  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Pln       .  5 1 . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Plxnb3    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ppp1r3g   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Prokr2    1  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ptafr     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ptgdr     .  . . . . . . . .  .  .  .  . . . .  .  . 1 . .  . . . . .  . . .  .
#> Ptger2    .  . . . . . . . .  .  1  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ptger3    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . 1  2
#> Ptger4    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ptgfr     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Pth1r     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . 1 .  . . .  .
#> Ptk7      .  . . . . 1 . . .  .  .  1  . . . .  1  1 . . .  . . . 1 .  . . .  .
#> Ret       .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Rho       .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ror1      .  . 1 . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Ror2      .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Rrh       .  . . . . . . . .  .  .  .  . . . 1  .  . . . .  . . . . .  . . .  .
#> Rspo3     .  . . . . . . . .  .  .  .  . . . .  1  . 2 . .  . 2 . . .  . . .  .
#> Rxfp3     .  . . . . . . . .  .  .  .  . . . .  .  . . . 1  . . . . .  . . .  .
#> Ryk       .  2 1 . . . 1 . 1  1  .  1  . . 1 .  .  1 1 . .  . . . 1 .  . . .  1
#> S1pr1     .  . . . . . . 4 .  .  .  .  . . 3 .  .  . . . .  . . . . .  5 1 .  .
#> S1pr2     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> S1pr3     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . 2 .  . . .  .
#> S1pr4     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> S1pr5     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Sctr      .  . . . . . . . .  .  .  1  . . . .  .  . . . .  . . . . .  . . .  .
#> Selplg    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Sema4d    .  . 1 1 . . . . 2  .  3  .  . . . .  1  1 . . .  . . . . .  . . .  .
#> Slc15a3   .  . . . 1 . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Slc17a6   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Slc17a7   9  . 9 4 4 . . 2 6 11  6  .  . . 1 2 11 16 . 3 4 12 . . . .  1 . .  .
#> Slc17a8   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Slc25a18  .  . 1 . . . . 1 .  1  .  3  1 . 1 1  .  . . . .  1 . . . .  . . .  .
#> Slc32a1   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Slc47a1   .  . . . . . . . .  .  .  .  . . . .  .  . 4 . .  . . . . .  . . .  .
#> Slco1a4   .  . . . . . 6 . .  .  .  .  . 2 . .  .  . . . .  . . . 1 .  . . .  .
#> Smo       1  1 . . . 2 . . 1  .  .  1  . . . .  1  . 2 . .  . 1 1 . .  . . .  3
#> Sox8      .  . . . . . . . 1  .  .  1  . . . .  .  . . . .  . . . . .  . . .  .
#> Sox9      .  . . . . . . 1 .  .  .  1  1 . 1 .  .  . . . .  2 . . . .  3 1 .  .
#> Sstr1     1  . 1 . . . . . .  .  .  .  . . . .  .  1 . . .  2 . . . .  . . .  .
#> Sstr2     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Sstr3     1  . . . . . . . .  .  1  .  . . . .  .  1 . . .  . . . . .  . . .  .
#> Sstr4     .  . . 1 1 . . . .  3  1  .  . . . .  1  1 7 . .  . . . . .  . . 1  1
#> Syt4      2  . 2 . . . . . 1  4  1  .  . . . .  2  2 . . .  4 . . . .  . . .  .
#> Taar2     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Taar3     .  1 . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Taar6     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Taar7e    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Tacr1     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  1
#> Tacr3     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Tas1r1    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Tas1r3    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Tas2r135  .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Tbxa2r    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . 1 .  . . .  .
#> Tek       .  . . . . . 2 . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Tie1      .  . . . . . 2 . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Timp4     .  . . . . . . 1 .  1  2  1  1 . . .  .  . . . .  . . . 1 .  1 1 .  .
#> Tjap1     .  . . . . 1 . . .  .  .  .  . . . 1  .  . . . .  . . . . .  . . .  .
#> Tmem108   1  . . 1 2 . . . .  2  2  .  . . . .  .  1 . . .  . . . . .  . . .  .
#> Traf4     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Trhr      1  . 2 . 1 . . . 1  1  3  .  . . . 2  4  1 . . .  . . . . .  . . .  .
#> Tyro3     .  . . . . . . . 1  1  .  .  . . . 2  .  . . . .  . . . . .  . . .  .
#> Uts2r     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> V1ra8     .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Vipr1     .  . 1 . . . . . .  1  .  .  . . . 1  1  . . . .  . . . . .  . . .  .
#> Vmn1r40   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Vmn1r43   .  . 1 . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Vmn1r46   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . 2 .  . . .  .
#> Vmn1r50   .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#> Vmn2r1    .  . . . . . . . .  .  .  .  . . . .  .  . . . .  . . . . .  . . .  .
#>                         
#> Abcc9      .  . . ......
#> Ackr1      .  . 1 ......
#> Ackr3      1  . . ......
#> Adcyap1r1  7  1 1 ......
#> Adgra1     1  2 1 ......
#> Adgra2     .  1 . ......
#> Adgra3     .  . . ......
#> Adgrb1     2  . . ......
#> Adgrb3     1  1 . ......
#> Adgre1     .  . . ......
#> Adgrf2     1  . . ......
#> Adgrf4     .  . . ......
#> Adgrf5     .  . . ......
#> Adgrg1     .  2 . ......
#> Adgrg2     .  . . ......
#> Adgrg4     1  . . ......
#> Adgrg6     .  . . ......
#> Adgrl1     2  2 1 ......
#> Adgrl2     .  1 . ......
#> Adgrl3     .  1 . ......
#> Adgrl4     .  5 . ......
#> Adgrv1     .  . . ......
#> Adora1     .  2 . ......
#> Adora2a    .  . . ......
#> Adora2b    .  . . ......
#> Adra1a     .  . . ......
#> Adra1b     .  . . ......
#> Adra1d     3  . . ......
#> Adra2a     .  1 . ......
#> Adra2b     .  . . ......
#> Adrb1      1  . . ......
#> Adrb2      .  . . ......
#> Agtr1a     .  . . ......
#> Aldh1l1    .  . . ......
#> Alk        .  . . ......
#> Amigo2     .  . . ......
#> Anxa11     2  . . ......
#> Aplnr      .  . . ......
#> Aqp4       .  1 . ......
#> Arhgap29   .  . . ......
#> Ascl1      .  . . ......
#> Atp13a5    .  . . ......
#> Avpr2      .  . . ......
#> Axl        .  . . ......
#> Baiap2     6  1 . ......
#> Bdkrb1     .  . . ......
#> Bdkrb2     .  . . ......
#> Blank-161  .  . . ......
#> Blank-155  .  1 . ......
#> Blank-152  .  . . ......
#> Blank-148  .  . . ......
#> Blank-145  .  . . ......
#> Blank-139  .  . . ......
#> Blank-128  .  . . ......
#> Blank-127  .  . . ......
#> Blank-104  .  . . ......
#> Blank-95   .  . . ......
#> Blank-83   .  . . ......
#> Blank-80   .  . . ......
#> Blank-69   .  . . ......
#> Blank-57   .  . . ......
#> Blank-45   .  . . ......
#> Brs3       .  . . ......
#> C1qa       .  . . ......
#> C1qb       .  . . ......
#> C3ar1      .  . . ......
#> C5ar1      .  . . ......
#> C5ar2      .  . . ......
#> Calcrl     .  . . ......
#> Casr       .  . . ......
#> Cckbr      .  . . ......
#> Ccr1       .  . . ......
#> Ccr2       .  . . ......
#> Ccr9       .  . . ......
#> Ccr10      .  . . ......
#> Cd300c2    .  . . ......
#> Celsr1     .  . . ......
#> Celsr2     7  2 . ......
#> Celsr3     1  . . ......
#> Cenpe      .  . . ......
#> Chrm1      6  . . ......
#> Chrm3      .  . . ......
#> Chrm4      .  . 1 ......
#> Chrm5      .  . . ......
#> Cldn5      . 12 . ......
#> Cmklr1     .  . . ......
#> Cnr1       .  . . ......
#> Crhr1      .  . . ......
#> Crhr2      1  . . ......
#> Csf1r      .  . . ......
#> Cspg5      4 14 . ......
#> Cx3cl1    16  . . ......
#> Cx3cr1     .  . . ......
#> Cxcl12     .  1 . ......
#> Cxcr4      .  . . ......
#> Cysltr2    .  . . ......
#> Ddr1       1  . . ......
#> Ddr2       .  . . ......
#> Dlk1       .  . . ......
#> Drd1       .  . . ......
#> Drd2       .  . . ......
#> Drd5       .  . . ......
#> Ednra      .  . . ......
#> Ednrb      .  1 . ......
#> Efemp1     .  . . ......
#> Egfr       .  2 . ......
#> Emcn       .  1 . ......
#> Eomes      .  . . ......
#> Epha1      .  . . ......
#> Epha2      .  . . ......
#> Epha3      .  . . ......
#> Epha4      6  1 . ......
#> Epha5      .  . . ......
#> Epha6      2  . 2 ......
#> Epha7      1  . 1 ......
#> Epha8      .  . . ......
#> Epha10     .  . . ......
#> Ephb1      .  . . ......
#> Ephb2      .  . . ......
#> Ephb3      .  1 1 ......
#> Ephb6      1  . 1 ......
#> Erbb2      .  1 . ......
#> Erbb3      .  . . ......
#> Erbb4      .  2 . ......
#> F2r        .  . . ......
#> F2rl3      .  . . ......
#> Fgfr2      .  . . ......
#> Fgfr3      .  3 . ......
#> Fgfrl1     .  1 . ......
#> Flt1       .  6 . ......
#> Flt3       .  . . ......
#> Flt4       .  . . ......
#> Fn1        .  1 . ......
#> Fpr1       .  . . ......
#> Fzd1       .  . . ......
#> Fzd2       .  . . ......
#> Fzd3       3  . . ......
#> Fzd4       .  3 . ......
#> Fzd5       .  . . ......
#> Fzd6       .  . . ......
#> Fzd7       .  . . ......
#> Fzd8       .  . . ......
#> Fzd9       .  . . ......
#> Fzd10      .  . . ......
#> Gabbr1     6  2 1 ......
#> Gabbr2     6  1 . ......
#> Gad1       3  . 1 ......
#> Galr1      .  . 1 ......
#> Gcgr       .  . . ......
#> Gfap       .  1 . ......
#> Ghsr       .  . . ......
#> Gipr       .  . . ......
#> Gjc3       .  . . ......
#> Glp2r      3  . 1 ......
#> Gpbar1     .  . . ......
#> Gper1      .  . . ......
#> Gpr1       .  . . ......
#> Gpr4       .  . . ......
#> Gpr6       .  . . ......
#> Gpr17      .  . . ......
#> Gpr20      .  . . ......
#> Gpr21      .  . . ......
#> Gpr22      2  . 2 ......
#> Gpr26      .  . . ......
#> Gpr27      2  . . ......
#> Gpr34      .  . . ......
#> Gpr37      .  . . ......
#> Gpr37l1    .  8 . ......
#> Gpr45      .  . . ......
#> Gpr52      .  1 . ......
#> Gpr55      .  . . ......
#> Gpr61      2  . . ......
#> Gpr62      1  . . ......
#> Gpr63      .  . . ......
#> Gpr75      .  . 1 ......
#> Gpr83      .  . . ......
#> Gpr85      2  . 1 ......
#> Gpr101     .  . . ......
#> Gpr135     1  . . ......
#> Gpr139     .  . . ......
#> Gpr146     .  . . ......
#> Gpr153     .  . . ......
#> Gpr156     .  . . ......
#> Gpr157     .  . 1 ......
#> Gpr158     .  . . ......
#> Gpr160     .  . . ......
#> Gpr161     .  . . ......
#> Gpr162     1  1 1 ......
#> Gpr173     .  . . ......
#> Gpr176     .  . 1 ......
#> Gpr182     .  . . ......
#> Gpr183     .  . . ......
#> Gprc5b     .  . 1 ......
#> Gprc5c     .  . . ......
#> Gramd3     .  1 . ......
#> Grin2b     7  . . ......
#> Grm1       .  . . ......
#> Grm2       1  . . ......
#> Grm3       1  . . ......
#> Grm4       .  . . ......
#> Grm5       8  . 2 ......
#> Grm7       2  . 2 ......
#> Grpr       .  . . ......
#> Hcar1      .  . . ......
#> Hcrtr2     .  . . ......
#> Hrh1       .  . . ......
#> Hrh2       .  . . ......
#> Hrh3       .  . . ......
#> Htr1a      .  . 1 ......
#> Htr1b      .  . . ......
#> Htr2a      .  . . ......
#> Htr4       1  . 1 ......
#> Htr5a      1  . . ......
#> Htr5b      .  . . ......
#> Htr6       .  . . ......
#> Htr7       .  . . ......
#> Igf1r      1  3 2 ......
#> Insr       2  . . ......
#> Insrr      .  . . ......
#> Kcnj8      .  . . ......
#> Kdr        .  1 . ......
#> Kiss1r     .  . . ......
#> Kit        1  . . ......
#> Lgr4       .  . . ......
#> Lgr6       .  . . ......
#> Lhcgr      .  . . ......
#> Lmod1      .  . . ......
#> Lmtk2      1  . . ......
#> Lpar1      .  . . ......
#> Lpar2      .  . . ......
#> Lpar4      .  . . ......
#> Lpar6      .  . . ......
#> Man1a      .  . . ......
#> Mas1       3  . 1 ......
#> Mc4r       .  . . ......
#> Mc5r       .  . . ......
#> Mchr1      .  . . ......
#> Mertk      .  3 . ......
#> Mlc1       1  3 . ......
#> Mrgpre     .  . . ......
#> Mrgprf     .  . . ......
#> Musk       .  . . ......
#> Myh11      .  . . ......
#> Nmbr       .  . . ......
#> Npy1r      .  . . ......
#> Npy2r      3  . 1 ......
#> Nrp1       3  1 . ......
#> Nrp2       .  . . ......
#> Ntrk2      .  5 . ......
#> Ntrk3      5  2 3 ......
#> Ntsr2      .  8 . ......
#> Olig1      .  1 . ......
#> Opn3       .  . . ......
#> Opn5       .  . . ......
#> Oprm1      1  . . ......
#> Oxtr       1  1 . ......
#> P2ry1      .  . . ......
#> P2ry2      .  . . ......
#> P2ry4      .  . . ......
#> P2ry6      .  . . ......
#> P2ry12     .  . . ......
#> P2ry14     1  . . ......
#> P2yr13     .  . . ......
#> Pcdh15     .  . . ......
#> Pdgfra     .  . . ......
#> Pdgfrb     .  . . ......
#> Peg10      .  . . ......
#> Pln        .  . . ......
#> Plxnb3     .  . . ......
#> Ppp1r3g    .  . . ......
#> Prokr2     .  . . ......
#> Ptafr      .  . . ......
#> Ptgdr      .  . . ......
#> Ptger2     .  . . ......
#> Ptger3     .  . . ......
#> Ptger4     .  . . ......
#> Ptgfr      .  . . ......
#> Pth1r      .  . . ......
#> Ptk7       .  . . ......
#> Ret        .  . . ......
#> Rho        .  . . ......
#> Ror1       .  . . ......
#> Ror2       .  . . ......
#> Rrh        .  . . ......
#> Rspo3      1  . . ......
#> Rxfp3      .  . . ......
#> Ryk        2  . . ......
#> S1pr1      .  5 . ......
#> S1pr2      .  . . ......
#> S1pr3      .  1 . ......
#> S1pr4      .  . . ......
#> S1pr5      .  . . ......
#> Sctr       .  . . ......
#> Selplg     .  . . ......
#> Sema4d     1  . . ......
#> Slc15a3    .  . . ......
#> Slc17a6    .  . . ......
#> Slc17a7   27  . 7 ......
#> Slc17a8    .  . . ......
#> Slc25a18   .  . . ......
#> Slc32a1    .  . . ......
#> Slc47a1    .  . . ......
#> Slco1a4    .  . . ......
#> Smo        2  . . ......
#> Sox8       .  . . ......
#> Sox9       .  1 . ......
#> Sstr1      5  . . ......
#> Sstr2      .  . . ......
#> Sstr3      .  . 1 ......
#> Sstr4      2  . . ......
#> Syt4       3  . . ......
#> Taar2      .  . . ......
#> Taar3      .  . . ......
#> Taar6      .  . . ......
#> Taar7e     .  . . ......
#> Tacr1      .  . . ......
#> Tacr3      .  . . ......
#> Tas1r1     .  . . ......
#> Tas1r3     .  . . ......
#> Tas2r135   .  1 . ......
#> Tbxa2r     .  . . ......
#> Tek        .  1 . ......
#> Tie1       .  . . ......
#> Timp4      .  1 . ......
#> Tjap1      .  . . ......
#> Tmem108    1  . . ......
#> Traf4      .  . . ......
#> Trhr       3  . . ......
#> Tyro3      .  1 1 ......
#> Uts2r      .  . . ......
#> V1ra8      .  . . ......
#> Vipr1      .  . . ......
#> Vmn1r40    .  . . ......
#> Vmn1r43    .  1 . ......
#> Vmn1r46    .  . . ......
#> Vmn1r50    .  . . ......
#> Vmn2r1     .  . . ......
#> 
#>  .....suppressing 429 columns in show(); maybe adjust options(max.print=, width=)
#>  ..............................