Set spatial locations

Function to set a spatial location slot

Usage

setSpatialLocations(
  gobject,
  x,
  spat_unit = NULL,
  name = "raw",
  provenance = NULL,
  verbose = TRUE,
  initialize = TRUE
)

Arguments

gobject

giotto object

x

spatLocsObj or list of spatLocsObj. Passing NULL will remove a specified set of spatial locations data.

spat_unit

spatial unit (e.g. "cell")

name

name of spatial locations, default "raw"

provenance

provenance information (optional)

verbose

be verbose

initialize

(default = FALSE) whether to initialize the gobject before returning

Value

giotto object

Details

Spatial information will be set to the nested location described by their tagged spat_unit and name information. An alternative location can also be specified through the respective params in this function.

See also