showGiottoSpatLocs

shows the available spatial locations

Usage

showGiottoSpatLocs(gobject, nrows = 4)

Arguments

gobject

giotto object

nrows

number of rows to print for each spatial location data.table

Value

prints the name and small subset of available data.table

See also

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> 3.2 read Giotto spatial centroid information
#> 3.3 read Giotto spatial overlap information
#> 4. read Giotto image information
#> python already initialized in this session
#>  active environment : 'giotto_env'
#>  python version : 3.10
#> checking default envname 'giotto_env'
#> a system default python environment was found
#> Using python path:
#>  "/usr/share/miniconda/envs/giotto_env/bin/python"

showGiottoSpatLocs(g)
#> └──Spatial unit "cell"
#>    └──S4 spatLocsObj "raw" coordinates:   (624 rows)
#>          An object of class spatLocsObj 
#>          provenance: cell
#>              ------------------------
#>             sdimx sdimy            cell_ID
#>             <int> <int>             <char>
#>          1:  5477 -4125 AAAGGGATGTAGCAAG-1
#>          2:  5959 -2808 AAATGGCATGTCTTGT-1
#>          3:  4720 -5202 AAATGGTCAATGTGCC-1
#>          4:  5202 -5322 AAATTAACGGGTAGCT-1
#>          
#>          ranges:
#>               sdimx sdimy
#>          [1,]  3069 -5442
#>          [2,]  6441 -2568
#>          
#>          
#>