Calculate features overlapped by polygons

Calculate subcellular points/feature info or image values overlapped by polygon annotations. This provides a summary of the spatial data overlapped by the polygon which can be further processed to become an expression matrix.

Usage

# S4 method for class 'giotto,missing'
calculateOverlap(
  x,
  name_overlap = NULL,
  spatial_info = NULL,
  feat_info = NULL,
  image_names = NULL,
  poly_subset_ids = NULL,
  return_gobject = TRUE,
  verbose = TRUE,
  ...
)

# S4 method for class 'giottoPolygon,giottoPoints'
calculateOverlap(
  x,
  y,
  name_overlap = NULL,
  poly_subset_ids = NULL,
  feat_subset_column = NULL,
  feat_subset_ids = NULL,
  count_info_column = NULL,
  return_gpolygon = TRUE,
  verbose = TRUE,
  ...
)

# S4 method for class 'giottoPolygon,giottoLargeImage'
calculateOverlap(
  x,
  y,
  name_overlap = NULL,
  poly_subset_ids = NULL,
  return_gpolygon = TRUE,
  verbose = TRUE,
  ...
)

# S4 method for class 'giottoPolygon,giottoAffineImage'
calculateOverlap(
  x,
  y,
  name_overlap = NULL,
  poly_subset_ids = NULL,
  return_gpolygon = TRUE,
  verbose = TRUE,
  ...
)

# S4 method for class 'giottoPolygon,SpatRaster'
calculateOverlap(
  x,
  y,
  name_overlap = NULL,
  poly_subset_ids = NULL,
  return_gpolygon = TRUE,
  verbose = TRUE,
  ...
)

# S4 method for class 'SpatVector,SpatRaster'
calculateOverlap(x, y, poly_subset_ids = NULL, verbose = TRUE, ...)

# S4 method for class 'SpatVector,SpatVector'
calculateOverlap(
  x,
  y,
  poly_subset_ids = NULL,
  feat_subset_column = NULL,
  feat_subset_ids = NULL,
  count_info_column = NULL,
  verbose = TRUE
)

Arguments

x

Object with spatial annotations: giottoPolygon, or SpatVector polygons. Can also be a giotto object

name_overlap

name for the overlap results (default to feat_info parameter)

spatial_info

character. Name polygon information

feat_info

character. Name of vector feature information to overlap

image_names

character vector. Name(s) of the image feature information to overlap

poly_subset_ids

character vector. (optional) Specific poly_IDs to use

return_gobject

return giotto object (default: TRUE)

verbose

be verbose

...

additional params to pass to methods.

y

Object with features to overlap: giottoPoints, giottoLargeImage, SpatVector points or SpatRaster

feat_subset_column

character. (optional) feature info attribute to subset feature points on when performing overlap calculation.

feat_subset_ids

(optional) values matched against in feat_subset_column in order to subset feature points when performing overlap calculation.

count_info_column

character. (optional) column with count information. Useful in cases when more than one detection is reported per point.

return_gpolygon

default = TRUE. Whether to return the entire giottoPolygon provided to x, but with the overlaps information appended or as a bare terra SpatVector

Value

Usually an object of the same class as x, with the overlaps information appended. return_* logical params usually allow return of a lower-level representation of the results instead. Only the SpatVector,SpatRaster method is different in that it returns a data.table

Details

feat_subset_column, feat_subset_ids, and count_info_column are specific to overlaps on feature points info, and should not be provided when overlapping image data. These three params can also be passed to the giotto method through the ... param when working with overlaps on feature points info.

Examples

g <- GiottoData::loadGiottoMini("vizgen")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> 3.2 read Giotto spatial centroid information
#> 3.3 read Giotto spatial overlap information
#> 4. read Giotto image information
#> python already initialized in this session
#>  active environment : 'giotto_env'
#>  python version : 3.10
gpoly <- getPolygonInfo(g,
    polygon_name = "aggregate",
    return_giottoPolygon = TRUE
)
gpoints <- getFeatureInfo(g, return_giottoPoints = TRUE)
gimg <- getGiottoImage(g, image_type = "largeImage")

slot(gpoly, "overlaps") <- NULL
overlaps(gpoly) # Should now be NULL
#> NULL

# detections from 2 z-layers are provided
table(gpoints$global_z)
#> 
#>     0     1 
#> 35889 43872 

# calculate all transcripts overlapped
out_all <- calculateOverlap(gpoly, gpoints)
#> 1. convert polygon to raster
#> 2. overlap raster and points
#> 3. add polygon information
#> 4. add points information
#> 5. create overlap polygon
#>  information
overlaps_all <- overlaps(out_all)
overlaps_all$rna
#>  class       : SpatVector 
#>  geometry    : points 
#>  dimensions  : 79761, 3  (geometries, attributes)
#>  extent      : 6400.037, 6900.032, -5149.983, -4699.979  (xmin, xmax, ymin, ymax)
#>  coord. ref. :  
#>  names       : poly_ID feat_ID feat_ID_uniq
#>  type        :   <chr>   <chr>        <num>
#>  values      :      NA    Mlc1            1
#>                     NA  Gprc5b            2
#>                     NA    Gfap            3

# calculate z1 only
out_z1 <- calculateOverlap(gpoly, gpoints,
    feat_subset_column = "global_z",
    feat_subset_ids = c(1)
)
#> 1. convert polygon to raster
#> 2. overlap raster and points
#> 3. add polygon information
#> 4. add points information
#> 5. create overlap polygon
#>  information
overlaps_z1 <- overlaps(out_z1)
overlaps_z1$rna
#>  class       : SpatVector 
#>  geometry    : points 
#>  dimensions  : 43872, 3  (geometries, attributes)
#>  extent      : 6400.043, 6900.03, -5149.983, -4699.984  (xmin, xmax, ymin, ymax)
#>  coord. ref. :  
#>  names       : poly_ID feat_ID feat_ID_uniq
#>  type        :   <chr>   <chr>        <num>
#>  values      :      NA  Gabbr2     3.62e+04
#>                     NA    Gfap     3.62e+04
#>                     NA    Gfap     3.62e+04

# overlap image to get sum intensities per cell
out_img <- calculateOverlap(gpoly, gimg, progress = FALSE)
#> Start image extract
#> End image extract
overlaps_img <- overlaps(out_img)
overlaps_img$intensity
#> $dapi_z0
#>                                          poly_ID mini_dataset_dapi_z0
#>                                           <char>                <num>
#>       1: 100210519278873141813371229408401071444                   18
#>       2: 100210519278873141813371229408401071444                   18
#>       3: 100210519278873141813371229408401071444                   18
#>       4: 100210519278873141813371229408401071444                   18
#>       5: 100210519278873141813371229408401071444                   18
#>      ---                                                             
#> 3170620:  98561957902191275233320065611022298397                   98
#> 3170621:  98561957902191275233320065611022298397                   98
#> 3170622:  98561957902191275233320065611022298397                   88
#> 3170623:  98561957902191275233320065611022298397                   88
#> 3170624:  98561957902191275233320065611022298397                   88
#> 

# giotto method
# calculate z0 overlaps and return as gobject
out_g <- calculateOverlap(g,
    feat_subset_column = "global_z",
    feat_subset_ids = 0
)
#> 1. convert polygon to raster
#> 2. overlap raster and points
#> 3. add polygon information
#> 4. add points information
#> 5. create overlap polygon
#>  information
overlaps(getPolygonInfo(out_g, return_giottoPolygon = TRUE))
#> $rna
#>  class       : SpatVector 
#>  geometry    : points 
#>  dimensions  : 35889, 3  (geometries, attributes)
#>  extent      : 6400.037, 6900.032, -5149.983, -4699.979  (xmin, xmax, ymin, ymax)
#>  coord. ref. :  
#>  names       : poly_ID feat_ID feat_ID_uniq
#>  type        :   <chr>   <chr>        <num>
#>  values      :      NA    Mlc1            1
#>                     NA  Gprc5b            2
#>                     NA    Gfap            3
#> 

# note that z0 and z1 nrows match that from the table of global z values.
# With points overlaps, all points are returned, but non-overlapped points
# only have an `NA` value for the `poly_ID` column. Overlapped points will
# have the `poly_ID` of their overlapping polygon.