Calculate features overlapped by polygons

Calculate features overlapped by polygons

Usage

# S4 method for class 'giotto,missing'
calculateOverlap(
  x,
  name_overlap = NULL,
  spatial_info = NULL,
  feat_info = NULL,
  image_names = NULL,
  poly_subset_ids = NULL,
  feat_subset_column = NULL,
  feat_subset_ids = NULL,
  count_info_column = NULL,
  return_gobject = TRUE,
  verbose = TRUE,
  ...
)

# S4 method for class 'giottoPolygon,giottoPoints'
calculateOverlap(
  x,
  y,
  name_overlap = NULL,
  poly_subset_ids = NULL,
  feat_subset_column = NULL,
  feat_subset_ids = NULL,
  count_info_column = NULL,
  return_gpolygon = TRUE,
  verbose = TRUE,
  ...
)

# S4 method for class 'giottoPolygon,SpatRaster'
calculateOverlap(
  x,
  y,
  name_overlap = NULL,
  poly_subset_ids = NULL,
  return_gpolygon = TRUE,
  verbose = TRUE,
  ...
)

# S4 method for class 'SpatVector,SpatRaster'
calculateOverlap(x, y, poly_subset_ids = NULL, verbose = TRUE, ...)

# S4 method for class 'SpatVector,SpatVector'
calculateOverlap(
  x,
  y,
  poly_subset_ids = NULL,
  feat_subset_column = NULL,
  feat_subset_ids = NULL,
  count_info_column = NULL,
  verbose = TRUE
)

Arguments

x

Object with spatial annotations: giottoPolygon, or SpatVector polygons. Can also be a giotto object

name_overlap

name for the overlap results (default to feat_info parameter)

spatial_info

character. Name polygon information

feat_info

character. Name of vector feature information to overlap

image_names

character vector. Name(s) of the image feature information to overlap

poly_subset_ids

(optional) character vector of poly_IDs to use

feat_subset_column

feature info column to subset features with

feat_subset_ids

ids within feature info column to use for subsetting

count_info_column

column with count information (optional)

return_gobject

return giotto object (default: TRUE)

verbose

be verbose

...

additional params to pass to methods

y

Object with features to overlap: giottoPoints, giottoLargeImage, SpatVector points or SpatRaster

return_gpolygon

default = TRUE. Whether to return the entire giottoPolygon provided to x, but with the overlaps information appended or as a bare terra SpatVector

Examples

if (FALSE) { # \dontrun{
x <- GiottoData::loadSubObjectMini("giottoPolygon")
y <- GiottoData::loadSubObjectMini("giottoPoints")
x@overlaps <- NULL
overlaps(x) # Should now be NULL

a <- calculateOverlap(x, y)
overlaps(a)
} # }