Subset Giotto

Internal to subset Giotto object including previous analyses for a specific spatial unit and feature type.
This is the master workflow for aggregate data subsetting. subsetGiotto, subsetGiottoLocs, and subsetGiottoLocsMulti only provide the params to pass to this function. This function also hides confusing but needed internal params from the exported functions.

The x_max, x_min, y_max, y_min params are only provided for subsetting of the feature info downstream of subsetGiottoLocs.

About polygon subsetting:
Polygons to subset are only subset using their poly_ID values which MUST match cell_ids, otherwise incorrect or empty polygons can be returned.

For subsetting the subcellular information only without editing the aggregate information, use subsetGiottoLocsSubcellular

Usage

.subset_giotto(
  gobject,
  spat_unit = ":all:",
  feat_type = "rna",
  spat_unit_fsub = ":all:",
  feat_type_ssub = ":all:",
  cell_ids = NULL,
  feat_ids = NULL,
  poly_info = spat_unit,
  x_max = NULL,
  x_min = NULL,
  y_max = NULL,
  y_min = NULL,
  verbose = FALSE,
  toplevel_params = 2
)

Arguments

gobject

giotto object

spat_unit, feat_type

character vector. (default = ':all:') One or more spatial units or feature types to subset. Accepts ':all:' as a token to subset across all available.

spat_unit_fsub

character vector. (default = ':all:') limit feat_id subsets to these spat_units

feat_type_ssub

character vector. (default = ':all:') limit cell_id subsets to these feat_types

cell_ids

character. cell IDs to keep

feat_ids

character. feature IDs to keep

poly_info

character. polygon info(s) to subset if present. (defaults to be the same as the spat_unit)

x_max, x_min, y_max, y_min

For feature info only. minimum and maximum x and y coordinates to keep

verbose

be verbose

toplevel_params

expected stackframe from which parameters should be extracted

Value

giotto object