Get polygon info

Get giotto polygon spatVector

Usage

getPolygonInfo(
  gobject = NULL,
  polygon_name = NULL,
  polygon_overlap = NULL,
  return_giottoPolygon = FALSE,
  verbose = TRUE,
  simplify = TRUE
)

Arguments

gobject

giotto object

polygon_name

name of polygons. Default is "cell"

polygon_overlap

include polygon overlap information

return_giottoPolygon

(Defaults to FALSE) Return as giottoPolygon S4 object

verbose

be verbose

simplify

logical. Whether or not to take object out of a list when there is a length of 1.

Value

spatVector

See also

Examples

g <- GiottoData::loadGiottoMini("vizgen")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> 3.2 read Giotto spatial centroid information
#> 3.3 read Giotto spatial overlap information
#> 4. read Giotto image information
#> python already initialized in this session
#>  active environment : 'giotto_env'
#>  python version : 3.10
#> checking default envname 'giotto_env'
#> a system default python environment was found
#> Using python path:
#>  "/usr/share/miniconda/envs/giotto_env/bin/python"

getPolygonInfo(g)
#>  class       : SpatVector 
#>  geometry    : polygons 
#>  dimensions  : 498, 1  (geometries, attributes)
#>  extent      : 6399.244, 6903.243, -5152.39, -4694.868  (xmin, xmax, ymin, ymax)
#>  source      : z0_spatInfo_spatVector.shp
#>  coord. ref. :  
#>  names       :                                 poly_ID
#>  type        :                                   <chr>
#>  values      :  40951783403982682273285375368232495429
#>                240649020551054330404932383065726870513
#>                274176126496863898679934791272921588227