overlapToMatrix

create a count matrix based on overlap results from calculateOverlap

Usage

# S4 method for class 'giotto'
overlapToMatrix(
  x,
  name = "raw",
  poly_info = NULL,
  feat_info = NULL,
  type = c("point", "intensity"),
  count_info_column = NULL,
  aggr_function = "sum",
  return_gobject = TRUE,
  verbose = TRUE,
  ...
)

# S4 method for class 'giottoPolygon'
overlapToMatrix(
  x,
  feat_info = "rna",
  type = c("point", "intensity"),
  count_info_column = NULL,
  output = c("Matrix", "data.table"),
  ...
)

# S4 method for class 'SpatVector'
overlapToMatrix(
  x,
  col_names = NULL,
  row_names = NULL,
  count_info_column = NULL,
  output = c("Matrix", "data.table"),
  verbose = TRUE,
  ...
)

# S4 method for class 'data.table'
overlapToMatrix(x, aggr_function = "sum", output = c("Matrix", "data.table"))

Arguments

x

object containing overlaps info. Can be giotto object or SpatVector points or data.table of overlaps generated from calculateOverlap

name

name for the overlap count matrix

poly_info

character. Polygon information to use

feat_info

character. Feature information to use

type

character. Type of overlap data (either 'point' or 'intensity')

count_info_column

column with count information

aggr_function

function to aggregate image information (default = sum)

return_gobject

return giotto object (default: TRUE)

verbose

be verbose

...

additional params to pass to methods

output

data format/class to return the results as

col_names, row_names

character vector. (optional) Set of row and col names that are expected to exist. This fixes the dimensions of the matrix since the overlaps information does not directly report rows and cols where no values were detected.

Value

giotto object or count matrix

Examples

g <- GiottoData::loadGiottoMini("vizgen")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> 3.2 read Giotto spatial centroid information
#> 3.3 read Giotto spatial overlap information
#> 4. read Giotto image information
#> python already initialized in this session
#>  active environment : 'giotto_env'
#>  python version : 3.10
#> checking default envname 'giotto_env'
#> a system default python environment was found
#> Using python path:
#>  "/usr/share/miniconda/envs/giotto_env/bin/python"
gpoly <- getPolygonInfo(g,
    polygon_name = "aggregate",
    return_giottoPolygon = TRUE
)
gpoints <- getFeatureInfo(g, return_giottoPoints = TRUE)

# calculate all transcripts overlapped
out_all <- calculateOverlap(gpoly, gpoints)
#> 1. convert polygon to raster
#> 2. overlap raster and points
#> 3. add polygon information
#> 4. add points information
#> 5. create overlap polygon
#>  information

overlapToMatrix(out_all)
#> Warning: [overlapToMatrix] expected col_names and row_names
#>  not provided together. Points aggregation Matrix output
#>  may be missing some cols and rows where no detections
#>  were found.
#> 337 x 462 sparse Matrix of class "dgCMatrix"
#>   [[ suppressing 33 column names ‘100210519278873141813371229408401071444’, ‘101161259912191124732236989250178928032’, ‘101488859781016188084173008420811094152’ ... ]]
#>                                                                                
#> Abcc9      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ackr1      .  . . 3 . .  . 1  .  1  7  . 2 2 1 .  .  . .  .  . . 1  . . 5 . . .
#> Ackr3      .  . 2 . . .  . .  .  1  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adcyap1r1  .  . 1 4 1 4  5 4  1  1  6  7 2 1 2 1  1  2 .  2  2 2 1  3 1 1 2 3 .
#> Adgra1     .  . 1 1 1 .  . 1  .  .  .  . 1 . . .  .  . .  2  . . 1  . . . . . .
#> Adgra2     1  . . . . .  . .  .  .  .  . . . . .  .  . 2  .  . . .  . . . . . .
#> Adgra3     .  1 . . . .  2 1  1  .  .  2 . . 2 .  .  . .  .  . . .  . . . . . .
#> Adgrb1     .  . . 3 . 1  . 3  .  6  6  1 2 . . 1  1  3 .  1  . 3 5  1 1 . 1 . .
#> Adgrb3     .  . . . . 3  . 2  1  1  3  1 . 1 . .  .  . .  .  . . .  2 . 2 3 1 .
#> Adgre1     .  . . . . .  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . 1 . .
#> Adgrf2     .  . . . . 1  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . . . .
#> Adgrf4     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adgrf5     4  . . . 1 .  . .  1  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adgrg1     2  . . . 1 .  4 .  2  .  .  . . . . .  1  . .  1  . . .  . . . 1 . .
#> Adgrg2     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adgrg4     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adgrg6     .  . . . . .  1 .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adgrl1     1  . 1 1 3 1  . 4  1  .  1  6 3 2 2 1  .  1 .  .  1 . 1  1 . 2 . . .
#> Adgrl2     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adgrl3     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adgrl4     2  . . . 1 .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adgrv1     .  . . . . .  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . 1 . .
#> Adora1     .  1 1 1 . .  . .  1  1  .  1 . . 1 .  4  2 .  .  1 . 1  . 1 . . 1 .
#> Adora2a    .  . . . . .  . .  .  .  .  . . . . .  1  . .  .  . . .  . . . . . .
#> Adora2b    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adra1a     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adra1b     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adra1d     .  . . . . .  . 1  .  1  1  . . 1 . .  .  2 .  .  . . 2  1 . 1 . . .
#> Adra2a     .  . . . . .  . 1  .  .  .  . . . . 1  .  . .  .  . . .  1 . . . . .
#> Adra2b     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Adrb1      .  . . . . .  . .  .  1  2  . . . . .  .  . .  .  . . .  . . . . . .
#> Adrb2      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Agtr1a     .  . . . . .  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . . . .
#> Aldh1l1    .  . . . 1 .  1 .  1  1  .  . . . . .  .  . .  2  . . .  . . . . . .
#> Alk        .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Amigo2     1  . . . . .  . .  .  .  .  1 . . . .  .  . .  .  . . .  1 . . . . .
#> Anxa11     .  . . 1 . 1  . 1  .  1  2  . 1 . 1 .  .  . 1  .  1 . 1  . 2 . . . .
#> Aplnr      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Aqp4       .  1 . . . .  3 .  4  .  .  . . . . .  1  1 . 11  . . .  . . . 1 . .
#> Arhgap29   2  . . . 1 .  . .  .  .  .  . . . . .  1  . .  .  . . .  . . . . . .
#> Ascl1      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Atp13a5    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Avpr2      .  . . . . .  . 1  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Axl        .  . . . . .  1 .  2  .  .  . . . . .  .  . 3  .  . . .  . . . . . .
#> Baiap2     .  1 5 2 1 2  . 3  .  4  7  1 4 1 1 .  .  2 .  1  1 5 2  5 . . . 1 .
#> Bdkrb1     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Bdkrb2     .  . . . . .  . .  .  .  .  . . . . .  .  . 1  .  . . .  . . . . . .
#> Blank-104  .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Blank-127  .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Blank-128  .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Blank-139  .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Blank-145  .  . . . . .  . .  .  1  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Blank-148  .  . . . . .  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . . . .
#> Blank-152  .  . . . . .  . .  .  .  .  . 1 . . .  .  . .  .  . . .  . . . . . .
#> Blank-155  .  . . 1 . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Blank-161  .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Blank-45   .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . 1 . . .
#> Blank-57   .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Blank-69   .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Blank-80   .  . . . . .  . 1  .  .  .  . . . . 1  .  . .  .  . . .  . . . . . .
#> Blank-83   .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Blank-95   .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Brs3       .  1 . . . .  . .  1  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> C1qa       .  1 . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . 3
#> C1qb       .  3 . . . .  . .  .  .  .  . . . . .  .  . .  .  . . 1  . . . . . 4
#> C3ar1      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> C5ar1      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> C5ar2      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . 1
#> Calcrl     .  . . . . .  . .  1  .  .  . . . . .  .  . 2  .  . . .  . . . . . .
#> Casr       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . 1 . . .
#> Cckbr      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ccr1       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ccr10      .  . . . . .  . .  .  .  .  1 . . . .  .  . .  .  . . .  . . . . . .
#> Ccr2       .  1 . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ccr9       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Cd300c2    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Celsr1     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . 1 . .
#> Celsr2     .  1 . 5 1 2  1 3  .  2  3  2 3 1 2 1  .  4 .  .  . . 2  3 . 1 1 1 .
#> Celsr3     .  . . . . .  . 1  .  .  .  . 1 1 . .  .  . .  4  . 1 1  . . . . . .
#> Cenpe      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Chrm1      .  . . 6 . 1  . 2  .  1  .  2 1 1 1 1  .  1 .  .  1 1 3  3 1 2 . . .
#> Chrm3      .  . . . . .  . 1  .  1  .  1 . . . .  .  . .  .  . . .  . . . . . .
#> Chrm4      .  . 2 . . .  . .  .  .  .  . . 1 . .  .  . .  .  . . 1  . . . . 1 .
#> Chrm5      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Cldn5      5  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Cmklr1     .  . . . . .  1 .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Cnr1       .  . . . . .  . .  .  1  .  1 . . . .  .  . .  .  1 . .  . . . . 2 1
#> Crhr1      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Crhr2      .  . . . . .  . 1  .  .  1  . . . . .  .  . .  .  . . .  . . . . . .
#> Csf1r      .  . . . . .  . .  .  .  .  . . . 1 .  .  . .  .  . . 1  . . . . . 7
#> Cspg5      .  . . . 5 1 12 .  9  .  3  2 1 1 . 1  .  . .  3  1 2 2  . . . 7 . .
#> Cx3cl1     .  . 1 5 1 .  1 5  1  2  2  7 2 4 . 3  .  3 1  2  3 5 6  6 1 4 . 3 .
#> Cx3cr1     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . 3
#> Cxcl12     3  . . . 2 .  . .  .  .  1  . 1 . . .  .  3 9  4  . . 1  . 1 3 . 1 .
#> Cxcr4      .  . . . 1 .  . .  .  .  .  . . . . .  .  . .  3  . . .  . . . . . .
#> Cysltr2    1  . . . . .  . 1  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ddr1       .  6 . . . 1  . .  3  .  2  . . 1 . .  2  . .  3  . . 1  . . . . . .
#> Ddr2       .  . 1 . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Dlk1       .  . . . . .  . .  .  .  .  . . . . .  4  . .  .  . . .  . . . 1 1 .
#> Drd1       .  . . . . .  . .  .  .  .  . . . . .  1  . .  .  . . .  . . . . . .
#> Drd2       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Drd5       .  . . . . .  . .  .  .  .  1 . . . .  .  . .  .  . . .  . . . . . .
#> Ednra      .  . . . . .  . .  1  .  .  . . . . .  .  . .  1  . . .  1 . . . . .
#> Ednrb      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Efemp1     .  . . . . .  . .  2  .  .  . . . . .  .  . 3  .  . . .  . . . . . .
#> Egfr       .  . . . . .  . .  1  .  .  . . . . .  .  . .  .  . . .  . . . . 1 .
#> Emcn       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Eomes      .  . . . . .  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . . . .
#> Epha1      .  . . . . .  . 1  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Epha10     .  1 1 1 . .  . .  .  .  2  . 2 . . .  .  . .  .  . . .  . . . . . .
#> Epha2      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Epha3      .  . . . . .  . .  .  .  .  . . . . .  .  . 1  .  . 1 .  1 . . . . .
#> Epha4      .  . 1 2 1 1  . 3  1  .  3  1 1 2 2 1  1  3 .  .  1 2 4  3 2 1 . 1 .
#> Epha5      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  1 . . . . .
#> Epha6      .  . 1 . . 1  . .  .  1  3  1 . 3 . 1  .  1 .  .  1 . 2  2 . . . 2 .
#> Epha7      .  . 1 . . 1  . 2  .  2  2  5 . 2 3 3  .  . 2  .  2 3 .  2 . . . 1 .
#> Epha8      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ephb1      .  2 . . 1 .  . 1  .  .  .  . . . . .  1  . .  .  . . .  . . . 2 . .
#> Ephb2      .  . . 1 . .  . .  .  1  .  . . . 1 .  .  . .  .  . . .  . . 1 . . .
#> Ephb3      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  1 . . . . .
#> Ephb6      .  . . . . .  . .  .  .  .  . . . . .  .  1 .  1  1 . 2  . . . . 1 1
#> Erbb2      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Erbb3      .  1 . . . .  . .  .  .  .  . . . . .  1  . .  .  . . .  . . . . . .
#> Erbb4      .  . . . . .  . .  .  .  .  . . . . .  1  . .  .  . . .  1 . . . . .
#> F2r        .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> F2rl3      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Fgfr2      .  6 . . . .  . .  1  .  .  . . . . .  1  . .  .  . . .  . . . . . .
#> Fgfr3      .  . . . 1 .  3 .  3  .  .  . . . . .  .  . .  1  . . .  . . . 1 . .
#> Fgfrl1     .  . . . 1 .  . .  4  .  .  . . . . .  .  . .  2  . . .  . . . . . .
#> Flt1      10  . . . 1 .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Flt3       .  . . . . .  . .  .  .  1  . . . . .  .  . .  .  . . .  1 . . . . .
#> Flt4       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Fn1        1  . . . 1 .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Fpr1       .  . . . . .  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . . . .
#> Fzd1       .  . . . . .  . .  2  .  .  . . . . .  .  . .  .  1 . .  . . . . . .
#> Fzd10      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Fzd2       .  . . . . .  1 .  .  .  .  . . . . .  .  . 1  .  . . .  . . . 1 . .
#> Fzd3       .  1 . 1 1 2  . 1  1  .  .  2 2 2 . 1  1  . 1  .  . 2 1  2 1 2 . 1 .
#> Fzd4       1  . . . . .  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . . . .
#> Fzd5       .  1 . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  1 . . . . .
#> Fzd6       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . 1 . .
#> Fzd7       .  . . . . .  . 1  .  1  .  . . . . .  .  . .  .  . . .  2 . . . . .
#> Fzd8       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . 1 1 . . .
#> Fzd9       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gabbr1     .  2 4 1 1 1  . 4  2  3 10  5 3 2 1 1  .  2 .  3  1 4 3  6 . 6 2 . 2
#> Gabbr2     .  . 3 2 . 1  . 1  .  .  3  1 3 1 3 2  .  2 .  .  2 2 2  1 1 5 . . 1
#> Gad1       .  . 2 1 . 2  . 3  1  1  .  . . . . 1  .  . .  .  1 . 2  2 . . . . .
#> Galr1      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . 1  . . . . . .
#> Gcgr       .  . . . . .  . .  .  .  1  . . . . .  .  . .  .  . . .  . . . . . .
#> Gfap       1 12 . . 6 .  . .  1  2  .  . 1 . . . 21  . . 13  . . .  . . . 1 . .
#> Ghsr       .  . . . . .  . .  .  .  .  . . . . 1  .  . .  .  . . .  1 . . . . .
#> Gipr       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gjc3       . 15 . . . .  . .  .  .  .  . . . . .  6  . .  .  . . .  . . . . . .
#> Glp2r      .  . . . . .  . .  .  2  5  2 2 . . .  .  1 .  .  . 2 1  2 1 . . 1 .
#> Gpbar1     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gper1      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr1       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr101     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr135     .  . . . . .  . .  .  .  .  . . . 1 .  .  . .  .  . 1 .  . . . . . .
#> Gpr139     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr146     .  . . . . .  . .  .  .  .  . 1 1 . .  .  . 1  .  . . .  . . . . . .
#> Gpr153     .  1 . . . .  . .  1  .  .  . . . . .  .  . .  .  . . .  . . . . . 1
#> Gpr156     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr157     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr158     .  . 4 . 1 2  . 1  .  .  1  . 1 . 2 .  .  . .  .  . . 1  1 . 1 . . .
#> Gpr160     1  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr161     .  . . . . .  . .  1  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr162     .  . 1 1 . 1  . 2  .  1  .  1 . 2 . .  .  3 .  1  . 2 .  . . . . . 1
#> Gpr17      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr173     .  . . . . .  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . . . .
#> Gpr176     .  . . . . .  . 1  .  .  1  . . . . .  .  1 .  .  . 1 .  . . . . 1 .
#> Gpr182     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr183     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr20      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr21      .  . . . . .  . .  .  1  .  . . . . .  2  . .  .  . . .  . . 1 . . 1
#> Gpr22      .  . . 2 1 .  . 2  .  1  1  . 2 1 . .  .  1 .  .  . 2 2  1 . 1 . 1 .
#> Gpr26      .  . 1 . . .  . .  .  2  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr27      .  . . . 1 .  . 1  .  .  .  1 . . . .  .  . .  1  1 . .  . . . . . .
#> Gpr34      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr37      .  . . . . .  . .  .  .  .  1 . . . .  .  . .  .  . . .  . . . . . .
#> Gpr37l1    .  . . . 1 . 10 .  9  .  .  . . . . .  .  . .  5  . 1 .  . . . 2 . .
#> Gpr4       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr45      .  . . . . .  . .  .  .  2  . . 1 . .  .  . .  .  . . .  . . . . . .
#> Gpr52      .  . . . . .  . 1  .  .  3  1 . . . .  .  . .  .  . . .  . . 1 . . .
#> Gpr55      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr6       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr61      .  . 1 1 . .  . 1  .  .  1  . . . . .  .  . .  .  . . .  . . . . 1 .
#> Gpr62      .  1 . . . .  . .  .  .  .  1 . . . .  1  . .  .  . . 1  1 . . . . .
#> Gpr63      .  . . . . .  1 .  .  .  2  . . . . .  .  1 .  .  . . .  . . . . 1 .
#> Gpr75      .  . . . 1 .  . .  .  .  1  . . . 1 .  .  . .  .  . 3 .  . . . . . .
#> Gpr83      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gpr85      .  . . . . 2  . 1  .  .  .  3 1 . . .  .  1 .  .  . 2 2  . . . . . .
#> Gprc5b     . 17 . . 2 .  1 1  4  .  5  . 1 . . .  7  1 .  1  . . 1  . . . 1 . .
#> Gprc5c     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Gramd3     .  . . . . .  . .  .  .  .  . . . . .  1  . .  .  . . .  . . . . . .
#> Grin2b     .  . 3 5 2 2  . 4  .  1  5  3 2 . 1 .  .  1 .  .  . 2 5  1 . 1 . 2 .
#> Grm1       .  . 1 . . .  . 2  .  2  4  1 . . 1 .  .  1 .  .  . . 1  . . 1 . . .
#> Grm2       .  . . 1 . .  . 1  .  2  2  2 1 1 . .  .  . .  1  . . 2  1 . 1 . . .
#> Grm3       1  6 . . . .  . .  2  .  .  . . . . .  3  . .  .  . . .  . . . 1 . .
#> Grm4       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Grm5       .  . 1 5 2 3  . 2  .  5  7  1 3 . 2 1  .  2 1  .  . 1 1  3 . 1 . . .
#> Grm7       .  . 3 1 1 2  . 2  .  3  4  4 2 . 1 .  1  2 .  .  2 . .  1 1 . . 2 .
#> Grpr       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Hcar1      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Hcrtr2     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Hrh1       .  . . . . .  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . . . .
#> Hrh2       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Hrh3       .  . . . 1 .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . 1 . . .
#> Htr1a      .  . . . 1 1  . 1  .  .  .  . . 1 . 1  .  . .  .  . . .  . . . . . .
#> Htr1b      .  . . . . .  . 1  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Htr2a      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Htr4       .  . . . . .  . .  .  .  3  1 . . . .  .  1 .  .  . . .  1 . . . . .
#> Htr5a      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Htr5b      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Htr6       .  . . . . .  . .  .  .  1  . . . . .  .  . .  .  . . .  . . . . . .
#> Htr7       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Igf1r      5  1 1 1 1 .  . 1  .  .  1  . . . . .  1  . 1  3  . . .  1 . . 1 1 1
#> Insr       .  1 1 . . .  . 1  .  2  3  1 3 . . 1  1  1 .  .  . . .  . 1 . . . .
#> Insrr      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Kcnj8      .  . . . 1 1  . .  .  1  .  1 . . . .  .  . .  .  . . .  . . . . . .
#> Kdr        4  . . . 1 .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Kiss1r     .  . . . . .  . .  .  .  .  . . . . 1  .  . .  .  . . .  . . . . . .
#> Kit        .  . . . 1 .  . .  .  .  .  . . . . .  .  . 2  2  . . .  . . . . . .
#> Lgr4       .  . . . 1 .  1 .  1  .  1  3 . . . .  .  . 3  1  . . .  . . . 1 1 .
#> Lgr6       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Lhcgr      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Lmod1      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Lmtk2      .  . 2 1 . .  . .  .  .  2  1 2 3 . 3  .  3 .  .  . . .  2 . . . . .
#> Lpar1      .  4 . . 1 .  . .  .  .  .  . . . . .  2  . 1  .  . . .  . . . . . .
#> Lpar2      .  . . . . .  . .  .  .  .  . . . . .  .  . 1  .  . . .  . . . . . .
#> Lpar4      .  . . . . .  . .  .  .  .  . . . . .  .  . 1  .  . . .  . . . . . .
#> Lpar6      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Man1a      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Mas1       .  . 2 2 . 3  . .  .  1  .  . . . 1 .  .  . .  .  2 . .  . . 1 . 1 .
#> Mc4r       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Mc5r       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Mchr1      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Mertk      .  . . . 2 .  2 .  3  .  .  . . . . .  .  . .  1  . . .  . . . 1 . 1
#> Mlc1       .  1 . . 1 .  2 .  .  .  .  . . . . .  .  . .  1  . . .  1 . . . . .
#> Mrgpre     .  . . . . .  . .  .  .  .  . . . 1 .  .  . .  .  . . .  . . . . . .
#> Mrgprf     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Musk       .  . . . . .  . .  .  .  .  . . . . .  .  . 1  .  . . .  . . . . . .
#> Myh11      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Nmbr       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Npy1r      .  . 1 . . .  . 1  .  .  1  . . 1 1 .  .  . 1  .  1 . .  . . 1 . 1 .
#> Npy2r      .  . . 1 . 2  . 2  .  1  1  1 . . 1 1  .  1 .  .  1 1 2  1 . 1 . 3 .
#> Nrp1       .  . . . . 1  . 2  3  .  .  2 1 . . 2  .  . .  1  1 . 1  1 1 . . . 3
#> Nrp2       .  1 1 . . .  . 1  .  1  .  . 1 . 2 .  .  . .  1  1 . .  . . 2 1 . .
#> Ntrk2      1  1 . 1 1 .  4 .  7  1  1  2 1 . . .  .  . .  9  . 1 .  . . . 2 . .
#> Ntrk3      .  . 3 5 1 4  . 3  .  6  4  3 1 2 3 2  .  3 .  .  4 4 3  4 1 2 . 1 .
#> Ntsr2      1  2 . . 1 .  3 . 10  .  .  . . . . .  1  . .  5  . . .  . . . 1 . .
#> Olig1      . 15 . . . 1  . .  1  .  .  . . . . .  3  1 .  .  . . .  . 1 . . . .
#> Opn3       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Opn5       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Oprm1      .  . . . . 1  . .  .  1  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Oxtr       .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> P2ry1      .  . . . . .  . .  .  .  .  . . . . .  .  . 1  .  . . .  . . . . . .
#> P2ry12     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . 2  . . . . . 4
#> P2ry14     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> P2ry2      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> P2ry4      .  . . . . .  . .  .  .  .  . . . 1 .  .  . .  .  . . .  . . . . . .
#> P2ry6      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> P2yr13     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . 1  . . . . . 3
#> Pcdh15     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Pdgfra     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Pdgfrb     .  . . . . .  . .  .  .  .  . . . . .  .  . 1  .  . . .  . . . . . .
#> Peg10      .  . . . . .  . 1  .  .  .  . . . . .  .  . .  .  . . .  . . . . . 1
#> Pln        .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Plxnb3     .  . . . . .  . .  .  .  .  . . . . .  2  . .  .  . . .  . . . . . .
#> Ppp1r3g    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Prokr2     .  . . . . .  . 1  .  .  .  . . . . .  .  . .  .  . . 1  1 . . . . .
#> Ptafr      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ptgdr      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ptger2     .  . 1 . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ptger3     .  . . . . .  . 1  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ptger4     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ptgfr      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Pth1r      .  . . . . .  . .  1  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ptk7       .  . . . . .  . .  .  1  1  . . . . .  .  . 1  1  . . .  . . . . . .
#> Ret        .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Rho        .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ror1       .  . . 1 . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ror2       .  . . . . .  . .  .  .  .  . . . . .  .  . 1  .  . . .  . . . . . .
#> Rrh        .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Rspo3      .  . 1 . . .  . .  .  .  .  . . . 1 1  .  1 .  .  2 . 1  . . . . . .
#> Rxfp3      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Ryk        .  . . 1 . .  . 1  1  .  3  1 . . 1 .  1  . .  1  . . .  . . . . . 2
#> S1pr1      1  . . . 1 .  3 .  3  .  .  . . . . .  .  . .  .  . . .  . . . 2 . .
#> S1pr2      .  1 . . . .  . .  .  .  .  . . . . .  .  . 1  .  . . .  . . . . . .
#> S1pr3      .  . . . . .  . 1  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> S1pr4      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> S1pr5      .  3 . . . .  . .  .  .  .  . . . . .  2  . .  .  . . .  . . . . . .
#> Sctr       .  . . . . .  . .  .  .  .  . . . . .  .  . .  1  . . .  . . . . . .
#> Selplg     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . 6
#> Sema4d     .  9 . 1 . .  . 1  .  .  2  2 2 . 1 .  .  . .  .  1 . .  . . 1 . 1 .
#> Slc15a3    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Slc17a6    .  . . . . .  . .  .  .  .  . . . . .  .  . .  5  . . .  . . . . . .
#> Slc17a7    .  . 7 7 4 4  3 7  . 10 25 11 8 7 7 4  . 11 .  1 10 7 7 11 2 5 . 5 .
#> Slc17a8    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Slc25a18   .  . . . . .  1 1  1  .  .  . . . . .  1  . .  2  . . 1  . . . 4 . .
#> Slc32a1    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Slc47a1    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Slco1a4    1  . . . . .  . .  .  .  .  . . . . .  .  . 1  .  . . .  . . . . . .
#> Smo        .  . . 1 . .  . .  .  1  .  . . . . .  .  . .  .  . . .  1 . . . . .
#> Sox8       .  3 . . . .  . 1  .  2  .  1 . 1 . .  2  . .  .  . . .  . 2 . . . .
#> Sox9       .  . . . 6 .  1 1  .  1  .  . . . . .  1  . .  1  . . .  . . . . . .
#> Sstr1      .  . . . . 1  . 3  .  2  .  1 3 . . 1  .  . .  .  1 2 .  2 . 1 . 1 .
#> Sstr2      .  . . . . .  . 1  .  .  2  . . . . .  .  . .  .  . . 1  . . 1 . 1 .
#> Sstr3      .  . . . . .  . .  .  .  1  . 1 . . .  .  . .  .  . 2 1  1 . . . . .
#> Sstr4      .  . . . . 1  1 1  .  .  .  1 1 3 2 .  .  1 .  .  . 1 1  . . . . . .
#> Syt4       .  . . 1 . .  . 1  .  .  .  . 1 . . .  .  1 .  .  . 1 1  2 . . . 1 .
#> Taar2      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Taar3      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . 1 .
#> Taar6      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . 1
#> Taar7e     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Tacr1      .  . . . . .  . .  .  .  .  . . . 1 .  .  . .  .  . . .  . . . . . .
#> Tacr3      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Tas1r1     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Tas1r3     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Tas2r135   .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Tbxa2r     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Tek        2  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Tie1       2  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Timp4      .  1 . . . .  1 .  1  .  .  . . . . .  .  . .  2  . . .  . . . . . .
#> Tjap1      .  . . . . 1  . .  .  .  .  . . . . .  1  . .  .  . . .  . . . . . .
#> Tmem108    .  . . 2 . 3  . 4  .  1  5  3 1 1 2 .  .  . .  .  . 2 2  1 1 . . . .
#> Traf4      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . 1
#> Trhr       .  . 2 1 . 1  . .  .  .  2  1 1 . 1 .  .  4 .  .  . . 1  1 . 1 . . .
#> Tyro3      .  9 . . . .  1 .  .  .  .  . . 1 1 .  .  . .  .  . . .  . . . . . .
#> Uts2r      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> V1ra8      .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Vipr1      .  . . . . .  . .  .  .  .  . 1 . . .  .  1 .  .  . . .  . . . . . .
#> Vmn1r40    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Vmn1r43    .  . . . . .  . .  .  .  .  1 . . . .  .  . .  .  . . .  . . . . . .
#> Vmn1r46    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Vmn1r50    .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#> Vmn2r1     .  . . . . .  . .  .  .  .  . . . . .  .  . .  .  . . .  . . . . . .
#>                           
#> Abcc9      .  . . . ......
#> Ackr1      .  . . . ......
#> Ackr3      .  . . . ......
#> Adcyap1r1  1  2 . 2 ......
#> Adgra1     .  2 . . ......
#> Adgra2     .  . . . ......
#> Adgra3     .  . 1 1 ......
#> Adgrb1     .  . . . ......
#> Adgrb3     3  1 . . ......
#> Adgre1     .  . . . ......
#> Adgrf2     .  . . . ......
#> Adgrf4     .  . . . ......
#> Adgrf5     .  . . . ......
#> Adgrg1     4  . . . ......
#> Adgrg2     .  . . . ......
#> Adgrg4     .  . . . ......
#> Adgrg6     .  . . . ......
#> Adgrl1     1  4 . . ......
#> Adgrl2     .  . . . ......
#> Adgrl3     .  . . . ......
#> Adgrl4     .  . . . ......
#> Adgrv1     .  . . . ......
#> Adora1     .  1 . 1 ......
#> Adora2a    .  . . . ......
#> Adora2b    .  . . . ......
#> Adra1a     .  . . . ......
#> Adra1b     .  . . . ......
#> Adra1d     .  1 . . ......
#> Adra2a     .  . . . ......
#> Adra2b     .  . . . ......
#> Adrb1      .  . . . ......
#> Adrb2      .  . . . ......
#> Agtr1a     .  1 . . ......
#> Aldh1l1    4  . . . ......
#> Alk        .  . . . ......
#> Amigo2     .  . . . ......
#> Anxa11     .  . . . ......
#> Aplnr      .  . . . ......
#> Aqp4       3  . . . ......
#> Arhgap29   .  1 . . ......
#> Ascl1      .  . . . ......
#> Atp13a5    1  . . . ......
#> Avpr2      .  . . . ......
#> Axl        2  . . . ......
#> Baiap2     .  1 . . ......
#> Bdkrb1     .  . . . ......
#> Bdkrb2     .  . . . ......
#> Blank-104  .  . . . ......
#> Blank-127  .  . . . ......
#> Blank-128  .  . . . ......
#> Blank-139  .  . . . ......
#> Blank-145  .  . . . ......
#> Blank-148  .  . . . ......
#> Blank-152  .  . . . ......
#> Blank-155  .  . . . ......
#> Blank-161  .  . . . ......
#> Blank-45   .  . . . ......
#> Blank-57   .  . . . ......
#> Blank-69   .  . . . ......
#> Blank-80   .  . . . ......
#> Blank-83   .  . . . ......
#> Blank-95   .  . . . ......
#> Brs3       .  . . . ......
#> C1qa       .  . 5 . ......
#> C1qb       .  . 6 . ......
#> C3ar1      .  . . . ......
#> C5ar1      .  . . . ......
#> C5ar2      .  . . . ......
#> Calcrl     .  . . . ......
#> Casr       .  . . . ......
#> Cckbr      .  . . . ......
#> Ccr1       .  . . . ......
#> Ccr10      .  . . . ......
#> Ccr2       .  . . . ......
#> Ccr9       .  . . . ......
#> Cd300c2    .  . . . ......
#> Celsr1     .  . . . ......
#> Celsr2     .  1 . . ......
#> Celsr3     .  1 . . ......
#> Cenpe      .  . . . ......
#> Chrm1      .  1 . . ......
#> Chrm3      .  . . . ......
#> Chrm4      .  . . . ......
#> Chrm5      .  . . . ......
#> Cldn5      .  . . . ......
#> Cmklr1     .  . . . ......
#> Cnr1       .  . . . ......
#> Crhr1      .  . . . ......
#> Crhr2      .  . . . ......
#> Csf1r      .  . 3 . ......
#> Cspg5      5  1 . . ......
#> Cx3cl1     .  3 . . ......
#> Cx3cr1     .  . 3 . ......
#> Cxcl12     .  . . . ......
#> Cxcr4      .  . . . ......
#> Cysltr2    .  . . . ......
#> Ddr1       .  2 . . ......
#> Ddr2       .  . . . ......
#> Dlk1       .  . . . ......
#> Drd1       .  . . . ......
#> Drd2       .  . . . ......
#> Drd5       .  . . . ......
#> Ednra      .  . . . ......
#> Ednrb      2  . . . ......
#> Efemp1     .  . . . ......
#> Egfr       .  1 . . ......
#> Emcn       .  . . . ......
#> Eomes      .  . . . ......
#> Epha1      .  . . . ......
#> Epha10     .  . . . ......
#> Epha2      .  . . . ......
#> Epha3      .  . . . ......
#> Epha4      .  1 . . ......
#> Epha5      .  . . . ......
#> Epha6      .  . . . ......
#> Epha7      .  1 . . ......
#> Epha8      .  . . . ......
#> Ephb1      .  1 . . ......
#> Ephb2      .  . . . ......
#> Ephb3      1  . . . ......
#> Ephb6      .  . 1 . ......
#> Erbb2      .  . . . ......
#> Erbb3      .  . . . ......
#> Erbb4      .  . . 1 ......
#> F2r        .  . . . ......
#> F2rl3      .  . . . ......
#> Fgfr2      1  . . . ......
#> Fgfr3      7  . 1 . ......
#> Fgfrl1     1  . . . ......
#> Flt1       .  1 . . ......
#> Flt3       .  1 . . ......
#> Flt4       .  . . . ......
#> Fn1        .  . . . ......
#> Fpr1       .  . . . ......
#> Fzd1       .  . . . ......
#> Fzd10      .  . . . ......
#> Fzd2       .  . . . ......
#> Fzd3       2  . . . ......
#> Fzd4       .  1 . . ......
#> Fzd5       .  . . . ......
#> Fzd6       .  . . . ......
#> Fzd7       .  . . . ......
#> Fzd8       .  . . . ......
#> Fzd9       .  . . . ......
#> Gabbr1     3  3 1 . ......
#> Gabbr2     .  2 . . ......
#> Gad1       1  . . . ......
#> Galr1      .  . . . ......
#> Gcgr       .  . . . ......
#> Gfap       1  . 2 2 ......
#> Ghsr       .  . . . ......
#> Gipr       .  . . . ......
#> Gjc3       .  . . . ......
#> Glp2r      .  . . . ......
#> Gpbar1     .  . . . ......
#> Gper1      .  . . . ......
#> Gpr1       .  . . . ......
#> Gpr101     .  . . . ......
#> Gpr135     .  . . . ......
#> Gpr139     .  . . . ......
#> Gpr146     1  . . . ......
#> Gpr153     .  . . . ......
#> Gpr156     .  . . . ......
#> Gpr157     .  . . . ......
#> Gpr158     .  . . . ......
#> Gpr160     .  . . . ......
#> Gpr161     .  . . . ......
#> Gpr162     1  . . . ......
#> Gpr17      .  . . . ......
#> Gpr173     .  . . . ......
#> Gpr176     .  . . . ......
#> Gpr182     .  . . . ......
#> Gpr183     .  . . . ......
#> Gpr20      .  . . . ......
#> Gpr21      .  . . . ......
#> Gpr22      .  . . . ......
#> Gpr26      .  . . . ......
#> Gpr27      .  . . . ......
#> Gpr34      .  . . . ......
#> Gpr37      .  . . . ......
#> Gpr37l1    6  . . . ......
#> Gpr4       .  . . . ......
#> Gpr45      .  1 . . ......
#> Gpr52      .  . . . ......
#> Gpr55      .  . . . ......
#> Gpr6       .  . . . ......
#> Gpr61      .  . . . ......
#> Gpr62      .  . . . ......
#> Gpr63      .  . . . ......
#> Gpr75      .  . . . ......
#> Gpr83      .  . . . ......
#> Gpr85      .  1 . . ......
#> Gprc5b     2  1 . . ......
#> Gprc5c     .  . . . ......
#> Gramd3     .  . . . ......
#> Grin2b     .  2 . . ......
#> Grm1       .  1 . . ......
#> Grm2       .  . . . ......
#> Grm3       4  . . . ......
#> Grm4       .  . . . ......
#> Grm5       .  2 . . ......
#> Grm7       .  3 . . ......
#> Grpr       .  . . . ......
#> Hcar1      .  . . . ......
#> Hcrtr2     .  . . . ......
#> Hrh1       .  . . . ......
#> Hrh2       .  . . . ......
#> Hrh3       .  . . . ......
#> Htr1a      .  . . . ......
#> Htr1b      .  . . . ......
#> Htr2a      .  . . . ......
#> Htr4       .  1 . . ......
#> Htr5a      .  . . . ......
#> Htr5b      .  . . . ......
#> Htr6       .  . . . ......
#> Htr7       .  . . . ......
#> Igf1r      2  1 . . ......
#> Insr       3  1 . . ......
#> Insrr      .  . . . ......
#> Kcnj8      .  . . . ......
#> Kdr        .  1 . . ......
#> Kiss1r     .  . . . ......
#> Kit        .  . . . ......
#> Lgr4       2  . 1 . ......
#> Lgr6       .  . . . ......
#> Lhcgr      .  . . . ......
#> Lmod1      .  . . . ......
#> Lmtk2      .  2 1 . ......
#> Lpar1      .  . . . ......
#> Lpar2      .  . . . ......
#> Lpar4      .  1 . . ......
#> Lpar6      .  . . . ......
#> Man1a      .  . . . ......
#> Mas1       .  . . . ......
#> Mc4r       .  . . . ......
#> Mc5r       .  . . . ......
#> Mchr1      .  . . . ......
#> Mertk      3  . . . ......
#> Mlc1       1  . . . ......
#> Mrgpre     .  . . . ......
#> Mrgprf     .  . . . ......
#> Musk       .  . . . ......
#> Myh11      .  . . . ......
#> Nmbr       .  . . . ......
#> Npy1r      .  . . . ......
#> Npy2r      .  1 . . ......
#> Nrp1       .  3 . . ......
#> Nrp2       .  . . . ......
#> Ntrk2      3  . . 1 ......
#> Ntrk3      .  . 1 . ......
#> Ntsr2     13  . . . ......
#> Olig1      .  . . . ......
#> Opn3       .  . . . ......
#> Opn5       .  . . . ......
#> Oprm1      .  . . . ......
#> Oxtr       .  . . . ......
#> P2ry1      .  . . . ......
#> P2ry12     .  . 5 . ......
#> P2ry14     .  . . . ......
#> P2ry2      .  . . . ......
#> P2ry4      .  . . . ......
#> P2ry6      .  . . . ......
#> P2yr13     .  . 1 1 ......
#> Pcdh15     .  . . . ......
#> Pdgfra     .  . . . ......
#> Pdgfrb     1  . . . ......
#> Peg10      .  . . . ......
#> Pln        .  . . . ......
#> Plxnb3     .  . . . ......
#> Ppp1r3g    .  . . . ......
#> Prokr2     .  . . . ......
#> Ptafr      .  . . . ......
#> Ptgdr      .  . . . ......
#> Ptger2     .  . . . ......
#> Ptger3     .  . . . ......
#> Ptger4     .  . . . ......
#> Ptgfr      .  . . . ......
#> Pth1r      .  . . . ......
#> Ptk7       .  1 . . ......
#> Ret        .  . . . ......
#> Rho        .  . . . ......
#> Ror1       .  . . . ......
#> Ror2       .  . . . ......
#> Rrh        .  . . . ......
#> Rspo3      .  . . . ......
#> Rxfp3      .  . . . ......
#> Ryk        .  1 . . ......
#> S1pr1      3  . . . ......
#> S1pr2      .  . . . ......
#> S1pr3      .  . . . ......
#> S1pr4      .  . . . ......
#> S1pr5      .  . . . ......
#> Sctr       .  . . . ......
#> Selplg     .  . 2 . ......
#> Sema4d     .  1 . . ......
#> Slc15a3    .  . 1 . ......
#> Slc17a6    .  . . . ......
#> Slc17a7    . 16 1 . ......
#> Slc17a8    .  . . . ......
#> Slc25a18   1  . . . ......
#> Slc32a1    .  . . . ......
#> Slc47a1    .  . . . ......
#> Slco1a4    .  . . . ......
#> Smo        .  . . . ......
#> Sox8       1  . . . ......
#> Sox9       1  . 2 1 ......
#> Sstr1      .  1 . . ......
#> Sstr2      .  . . . ......
#> Sstr3      .  1 . . ......
#> Sstr4      .  1 . . ......
#> Syt4       .  2 . . ......
#> Taar2      .  . . . ......
#> Taar3      .  . . . ......
#> Taar6      .  . . . ......
#> Taar7e     .  . . . ......
#> Tacr1      .  . . . ......
#> Tacr3      .  . . . ......
#> Tas1r1     .  . . . ......
#> Tas1r3     .  . . . ......
#> Tas2r135   .  . . . ......
#> Tbxa2r     .  . . . ......
#> Tek        .  . . . ......
#> Tie1       .  . . . ......
#> Timp4      .  . . . ......
#> Tjap1      .  . . . ......
#> Tmem108    .  1 . . ......
#> Traf4      .  . . . ......
#> Trhr       .  1 . . ......
#> Tyro3      .  . . . ......
#> Uts2r      .  . . . ......
#> V1ra8      .  . . . ......
#> Vipr1      .  . . . ......
#> Vmn1r40    .  . . . ......
#> Vmn1r43    .  . . . ......
#> Vmn1r46    .  . . . ......
#> Vmn1r50    .  . . . ......
#> Vmn2r1     .  . . . ......
#> 
#>  .....suppressing 429 columns in show(); maybe adjust options(max.print=, width=)
#>  ..............................